Gene-to-symbol mappings

Gene-to-symbol mappings

Gene2Symbol(object, ...)

# S4 method for DataFrame
Gene2Symbol(object, format = c("makeUnique", "1:1",
  "long"))

# S4 method for GRanges
Gene2Symbol(object, format = c("makeUnique", "1:1",
  "long"))

# S4 method for SummarizedExperiment
Gene2Symbol(object,
  format = c("makeUnique", "1:1", "long"))

Arguments

object
Object.

...
Additional arguments.

format
character(1). Formatting method to apply:

"makeUnique": Recommended. Apply make.unique() to the geneName column. Gene symbols are made unique, while the gene IDs remain unmodified.

"1:1": For gene symbols that map to multiple gene IDs, select only the first annotated gene ID.

"long": Return geneID and geneName columns unmodified in long format.

Value

Gene2Symbol. Contains a DataFrame with geneID and geneName columns.

Note

For some organisms, gene names and gene symbols do not map 1:1 (e.g. Homo sapiens and Mus musculus). Refer to the format argument here in the documentation for approaches that deal with this issue.

Genome metadata

We recommend slotting organism, genomeBuild, and ensemblRelease into metadata().

See also

Examples

data(rse, package = "acidtest")
x <- Gene2Symbol(rse)
print(x)

#> Gene2Symbol with 500 rows and 2 columns
#>                  geneID    geneName
#>             <character> <character>
#> gene001 ENSG00000000003      TSPAN6
#> gene002 ENSG00000000005        TNMD
#> gene003 ENSG00000000419        DPM1
#> gene004 ENSG00000000457       SCYL3
#> gene005 ENSG00000000460    C1orf112
#> ...                 ...         ...
#> gene496 ENSG00000032742       IFT88
#> gene497 ENSG00000033011        ALG1
#> gene498 ENSG00000033030      ZCCHC8
#> gene499 ENSG00000033050       ABCF2
#> gene500 ENSG00000033100       CHPF2