Access to the Data Inside an S4 Method Definition

Access to the Data Inside an S4 Method Definition

methodFunction(f, signature, package)

methodFormals(f, signature, package)

hasLocal(definition)

extractLocal(definition)

Arguments

f
a generic function or the character-string name of one.

signature
the signature of classes to match to the arguments of f. See the details below.

package
string. Package name.

definition
A function definition, which will become the method called when the arguments in a call to f match the classes in signature, directly or through inheritance. The definition must be a function with the same formal arguments as the generic; however, setMethod() will handle methods that add arguments, if … is a formal argument to the generic. See the Details section.

Value

methodFormals(): list. Extract the sealed formal arguments.

methodFunction(): function. Extract the sealed function.

See also

Examples

library(methods)
library(BiocGenerics)

#> Loading required package: parallel

#> 
#> Attaching package: 'BiocGenerics'

#> The following objects are masked from 'package:parallel':
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from 'package:base':
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colMeans,
#>     colnames, colSums, dirname, do.call, duplicated, eval, evalq,
#>     Filter, Find, get, grep, grepl, intersect, is.unsorted, lapply,
#>     lengths, Map, mapply, match, mget, order, paste, pmax, pmax.int,
#>     pmin, pmin.int, Position, rank, rbind, Reduce, rowMeans, rownames,
#>     rowSums, sapply, setdiff, sort, table, tapply, union, unique,
#>     unsplit, which, which.max, which.min

f <- "as.data.frame"
signature <- "ANY"

## Function.
x <- methodFunction(f = f, signature = signature)
class(x)

#> [1] "function"

formals(x)

#> $x
#> 
#> 
#> $row.names
#> NULL
#> 
#> $optional
#> [1] FALSE
#> 
#> $...
#> 
#> 

## Formals.
x <- methodFormals(f = f, signature = signature)
class(x)

#> [1] "pairlist"

print(x)

#> $x
#> 
#> 
#> $row.names
#> NULL
#> 
#> $optional
#> [1] FALSE
#> 
#> $...
#> 
#>