Aggregate Features

Aggregate Features

aggregateFeatures(object, ...)

# S4 method for dgCMatrix
aggregateFeatures(object, groupings)

# S4 method for matrix
aggregateFeatures(object, groupings)

Arguments

object
Object.

...
Additional arguments (for the S4 generic definition).

groupings
Feature groupings (e.g. gene or transcript IDs), defined as a named factor. The pooled features must be defined as the factor levels, and the original features as the names of the factor.

Value

Object of same class with aggregated features (rows).

See also

Other Data Management Utilities: aggregateReplicates

Examples

counts <- data.frame(
    "sample1" = c(0, 1, 2, 3),
    "sample2" = c(1, 2, 3, 4),
    row.names = c("gene1.1", "gene1.2", "gene2.1", "gene2.2")
)

groupings <- factor(c("gene1", "gene1", "gene2", "gene2"))
names(groupings) <- rownames(counts)

# matrix
mat <- as(counts, "matrix")
aggregateFeatures(mat, groupings = groupings)

#>       sample1 sample2
#> gene1       1       3
#> gene2       5       7

# dgCMatrix
dgc <- as(mat, "dgCMatrix")
aggregateFeatures(dgc, groupings = groupings)

#> 2 x 2 sparse Matrix of class "dgCMatrix"
#>       sample1 sample2
#> gene1       1       3
#> gene2       5       7