Ensembl Annotations

Quickly obtain gene annotations from Ensembl.

annotable(object, ...)

# S4 method for character
annotable(object, format = "gene", genomeBuild = NULL,
  release = NULL, uniqueSymbols = TRUE, quiet = FALSE)

# S4 method for data.frame
annotable(object)

Arguments

object
Object. Default recommended usage is to provide full latin organism name as a string. Also supports input of pre-built tibbles from the annotables data package.

...
Additional arguments (for the S4 generic definition).

format
Desired table format, either gene, tx2gene, or gene2symbol.

genomeBuild
Optional. Check to see if a specific genome build is supported.

release
Optional. Ensembl release version. Defaults to the most current release available on AnnotationHub.

uniqueSymbols
Make gene symbols unique. Recommended by default.

quiet
If TRUE, suppress any status messages and/or progress bars.

Value

data.frame with unique rows per gene or transcript.

Using AnnotationHub

The default recommended approach is to simply specify the desired organism, using the full latin name. For example, we can obtain human annotations with Homo sapiens and mouse annotations with Mus musculus. Under the hood, the function fetches these annotations from AnnotationHub using the ensembldb package. AnnotationHub supports versioned Ensembl releases, back to version 87.

Using Pre-Compiled Annotables

Generally we' don't recommend this approach. However, pass-in of a tibble generated using biomaRt in the annotables data package is useful for adding support for the outdated (but still frequently used) GRCh37/hg19 human genome.

Broad Class Definitions

A broadClass column is added, which generalizes the gene types into a smaller number of semantically-meaningful groups:

coding

noncoding

pseudo

small

decaying

ig (immunoglobulin)

tcr (T cell receptor)

other

See also

AnnotationHub.

ensembldb.

Other Gene Annotation Utilities: detectOrganism, gene2symbolFromGFF, gene2symbol, symbol2gene, tx2geneFromGFF, tx2gene

Examples

annotable("Homo sapiens") %>% glimpse()

#> Loading Ensembl annotations from AnnotationHub
#> 2017-10-27

#> EnsDB AH57757: Homo sapiens Ensembl 90

#> Observations: 64,661
#> Variables: 11
#> $ ensgene        <chr> "ENSG00000000003", "ENSG00000000005", "ENSG000000004...
#> $ symbol         <chr> "TSPAN6", "TNMD", "DPM1", "SCYL3", "C1orf112", "FGR"...
#> $ description    <chr> "tetraspanin 6 [Source:HGNC Symbol;Acc:HGNC:11858]",...
#> $ biotype        <chr> "protein_coding", "protein_coding", "protein_coding"...
#> $ broadClass     <chr> "coding", "coding", "coding", "coding", "coding", "c...
#> $ geneSeqStart   <int> 100627109, 100584802, 50934867, 169849631, 169662007...
#> $ geneSeqEnd     <int> 100639991, 100599885, 50958555, 169894267, 169854080...
#> $ seqName        <chr> "X", "X", "20", "1", "1", "1", "1", "6", "6", "6", "...
#> $ seqStrand      <int> -1, 1, -1, -1, 1, -1, 1, -1, -1, 1, -1, 1, -1, 1, 1,...
#> $ seqCoordSystem <chr> "chromosome", "chromosome", "chromosome", "chromosom...
#> $ entrez         <list> [7105, 64102, 8813, 57147, 55732, 2268, 3075, 2519,...

# Legacy GRCh37/hg19 genome build support
annotable("Homo sapiens", genomeBuild = "GRCh37") %>% glimpse()

#> Observations: 63,677
#> Variables: 10
#> $ ensgene     <chr> "ENSG00000000003", "ENSG00000000005", "ENSG00000000419"...
#> $ symbol      <chr> "TSPAN6", "TNMD", "DPM1", "SCYL3", "C1orf112", "FGR", "...
#> $ description <chr> "tetraspanin 6 [Source:HGNC Symbol;Acc:11858]", "tenomo...
#> $ biotype     <chr> "protein_coding", "protein_coding", "protein_coding", "...
#> $ broadClass  <chr> "coding", "coding", "coding", "coding", "coding", "codi...
#> $ chr         <chr> "X", "X", "20", "1", "1", "1", "1", "6", "6", "6", "1",...
#> $ start       <int> 99883667, 99839799, 49551404, 169818772, 169631245, 279...
#> $ end         <int> 99894988, 99854882, 49575092, 169863408, 169823221, 279...
#> $ strand      <int> -1, 1, -1, -1, 1, -1, 1, -1, -1, 1, -1, 1, -1, 1, 1, 1,...
#> $ entrez      <list> [7105, 64102, 8813, 57147, 55732, 2268, 3075, 2519, 27...

# Convert pre-compiled annotables tibble (advanced)

# NOT RUN {
annotable(annotables::grch38) %>% glimpse()
# }