Force an Object to Belong to a Class

Coerce an object to a given class.

as.DataFrame(x)

as.SummarizedExperiment(x)

Arguments

x
Object.

...
Additional arguments.

Value

Object of new class.

list

It is often useful to coerce an S4 object to a flat list for archival storage. Here we are providing the coerceS4ToList() function, which consistently coerces the slots in any S4 to a standard list. Additionally, here we have improved support for SummarizedExperiment to list coercion, returning the slots as a list.

tibble

Coerce an object to a tibble (tbl_df, data.frame) using either S3 or S4 methods:

S3: as_tibble(x) (or as_tibble() alias).

S4: as(object, Class = "tbl_df").

Tibbles don't support row name assignment, so here we are ensuring they are kept by moving them to a column named rowname by default. This helps avoid downstream unexpected data loss when using the dplyr chain of single table verbs, such as dplyr::arrange(), dplyr::filter(), or dplyr::mutate().

This behavior can be overriden in the S3 method by setting rowname = NULL instead, which is the current default in the tibble package. The S4 coercion method doesn't support arguments, and therefore always attempts to move rownames automatically, if defined.

Conversely, when coercing a tibble back to an S4 DataFrame, our as(tbl_df, Class = "DataFrame") method looks for the rowname column and will attempt to move it back to base::rownames() automatically, unless there are duplicates present.

References

Chambers, John M. (2016) Extending R, Chapman & Hall. (Chapters 9 and 10.)

See also

Examples

data(rse)

## DataFrame to tbl_df ====
data <- SummarizedExperiment::colData(rse)
print(data)

#> DataFrame with 12 rows and 1 column
#>          condition
#>           <factor>
#> sample01         A
#> sample02         A
#> sample03         A
#> sample04         A
#> sample05         A
#> ...            ...
#> sample08         B
#> sample09         B
#> sample10         B
#> sample11         B
#> sample12         B

x <- as(data, "tbl_df")
print(x)

#> # A tibble: 12 x 2
#>    rowname  condition
#>    <chr>    <fct>    
#>  1 sample01 A        
#>  2 sample02 A        
#>  3 sample03 A        
#>  4 sample04 A        
#>  5 sample05 A        
#>  6 sample06 A        
#>  7 sample07 B        
#>  8 sample08 B        
#>  9 sample09 B        
#> 10 sample10 B        
#> 11 sample11 B        
#> 12 sample12 B        

x <- tibble::as_tibble(data)
print(x)

#> # A tibble: 12 x 2
#>    rowname  condition
#>    <chr>    <fct>    
#>  1 sample01 A        
#>  2 sample02 A        
#>  3 sample03 A        
#>  4 sample04 A        
#>  5 sample05 A        
#>  6 sample06 A        
#>  7 sample07 B        
#>  8 sample08 B        
#>  9 sample09 B        
#> 10 sample10 B        
#> 11 sample11 B        
#> 12 sample12 B        

## GRanges to tbl_df ====
data <- SummarizedExperiment::rowRanges(rse)

x <- as(data, "tbl_df")

x <- tibble::as_tibble(data)
colnames(x)

#>  [1] "rowname"     "seqnames"    "start"       "end"         "width"      
#>  [6] "strand"      "geneID"      "geneName"    "geneBiotype" "broadClass" 
#> [11] "entrezID"   

## sparseMatrix to data.frame ====
data(sparse)

x <- BiocGenerics::as.data.frame(sparse)
class(x)

#> [1] "data.frame"

x <- as(sparse, "data.frame")
class(x)

#> [1] "data.frame"

## tbl_df to DataFrame ====
data <- tibble::as_tibble(datasets::iris)
x <- as(data, "DataFrame")