Assign and Save Data

Assigns a new object by name to the current working environment then saves the newly assigned object, specified by the "dir" argument.

assignAndSaveData(name, object, envir = parent.frame(),
  dir = getOption("basejump.save.dir", "."),
  ext = getOption("basejump.save.ext", "rds"),
  overwrite = getOption("basejump.save.overwrite", TRUE),
  compress = getOption("basejump.save.compress", TRUE))

Arguments

name
string. Desired variable name.

object
Object.

envir
environment. Environment to use for assignment. Defaults to parent.frame(), the calling environment.

dir
string. Directory path.

ext
string. R data serialized (RDS; "rds") or R data ("rda", "RData"). RDS is preferred when saving single objects per file, which is always the convention of saveData(), regardless of the extension used.

overwrite
boolean. Overwrite existing file.

compress
logical or character string specifying whether saving to a named file is to use compression. TRUE corresponds to gzip compression, and character strings "gzip", "bzip2" or "xz" specify the type of compression. Ignored when file is a connection and for workspace format version 1.

Value

Invisible named string. File path.

Note

This function attempts to follow the same order as assign().

Examples

data(rse)
assignAndSaveData(name = "example", object = rse, ext = "rds")

#> Saving example.rds to /data01/git/basejump/docs/reference

exists("example", inherits = FALSE)

#> [1] TRUE

file.exists("example.rds")

#> [1] TRUE

## Clean up.
rm(example)
unlink("example.rds")