Install or update Bioconductor packages

This function is a convenience wrapper for BiocInstaller::biocLite(). It sources the required biocLite.R file from the Bioconductor website automatically then calls biocLite().

biocLite(pkgs = NULL, ...)

Arguments

pkgs
Character vector of package names to install or update. Supports Bioconductor, CRAN, and GitHub repositories.

...
Passthrough parameters.

See also

This function is a modified version of rafalib::install_bioc().

Examples

# NOT RUN {
# Load Bioconductor and update packages
biocLite()

# Install packages
biocLite("limma")            # Bioconductor
biocLite("dplyr")            # CRAN
biocLite("tidyverse/dplyr")  # GitHub
# }