Collapse Utilities

Collapse Utilities

collapse(object, ...)

# S4 method for character
collapse(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

# S4 method for integer
collapse(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

# S4 method for data.frame
collapse(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

# S4 method for DataFrame
collapse(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

# S4 method for matrix
collapse(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

# S4 method for tbl_df
collapse(object, sep = ", ", unique = TRUE,
  sort = TRUE, keepNA = FALSE, ...)

Arguments

object
Object.

...
Additional arguments, passed to glue::collapse().

sep
Separator. Defaults to comma.

unique
Unique values.

sort
Sort values.

keepNA
Keep NA values.

Value

For vector: String.

For column data: Single row.

See also

base::toString().

glue::collapse().

Examples

# Designed primarily for `dims` objects
mtcars %>%
    head() %>%
    collapse() %>%
    t()

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("data.frame"), structure(function (x) standardGeneric("collapse"), generic = structure("collapse", package = "IRanges"), package = "IRanges", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x) stop("invalid call in method dispatch to 'collapse' (no default method)",     domain = NA))(x), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function 'collapse' for signature '"data.frame"'

# Vectors are also supported
groceries <- c("milk", "eggs", "eggs", "veggies", NA)
collapse(groceries)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("character"), structure(function (x) standardGeneric("collapse"), generic = structure("collapse", package = "IRanges"), package = "IRanges", group = list(), valueClass = character(0), signature = "x", default = `\001NULL\001`, skeleton = (function (x) stop("invalid call in method dispatch to 'collapse' (no default method)",     domain = NA))(x), class = structure("standardGeneric", package = "methods")),     <environment>): unable to find an inherited method for function 'collapse' for signature '"character"'

collapse(groceries, unique = FALSE, sort = FALSE)

#> Error in collapse(groceries, unique = FALSE, sort = FALSE): unused arguments (unique = FALSE, sort = FALSE)