Combining or merging different Bioconductor data structures

The combine generic function handles methods for combining or merging different Bioconductor data structures. It should, given an arbitrary number of arguments of the same class (possibly by inheritance), combine them into a single instance in a sensible way (some methods may only combine 2 objects, ignoring ... in the argument list; because Bioconductor data structures are complicated, check carefully that combine does as you intend).

# S4 method for SummarizedExperiment,SummarizedExperiment
combine(x, y)

# S4 method for SingleCellExperiment,SingleCellExperiment
combine(x, y)

Arguments

x
One of the values.

y
A second value.

Value

SummarizedExperiment.

Details

There are two basic combine strategies. One is an intersection strategy. The returned value should only have rows (or columns) that are found in all input data objects. The union strategy says that the return value will have all rows (or columns) found in any one of the input data objects (in which case some indication of what to use for missing values will need to be provided).

These functions and methods are currently under construction. Please let us know if there are features that you require.

Note

We're attempting to make this as strict as possible, requiring:

Rows (genes) across objects must be identical.

SummarizedExperiment::rowRanges() and/or SummarizedExperiment::rowData() metadata must be identical.

SummarizedExperiment::colData() must contain the same columns.

Specific metadata must be identical (see metadata parameter).

Methods

The following methods are defined in the BiocGenerics package:

combine(x=ANY, missing)

Return the first (x) argument unchanged.

combine(data.frame, data.frame)

Combines two data.frame objects so that the resulting data.frame contains all rows and columns of the original objects. Rows and columns in the returned value are unique, that is, a row or column represented in both arguments is represented only once in the result. To perform this operation, combine makes sure that data in shared rows and columns are identical in the two data.frames. Data differences in shared rows and columns usually cause an error. combine issues a warning when a column is a factor and the levels of the factor in the two data.frames are different.

combine(matrix, matrix)

Combined two matrix objects so that the resulting matrix contains all rows and columns of the original objects. Both matricies must have dimnames. Rows and columns in the returned value are unique, that is, a row or column represented in both arguments is represented only once in the result. To perform this operation, combine makes sure that data in shared rows and columns are all equal in the two matricies.

Additional combine methods are defined in the Biobase package for AnnotatedDataFrame, AssayData, MIAME, and eSet objects.

See also

Examples

data(rse, sce)

## SummarizedExperiment ====
x <- rse
colnames(x)

#>  [1] "sample01" "sample02" "sample03" "sample04" "sample05" "sample06"
#>  [7] "sample07" "sample08" "sample09" "sample10" "sample11" "sample12"

SummarizedExperiment::colData(x)

#> DataFrame with 12 rows and 1 column
#>          condition
#>           <factor>
#> sample01         A
#> sample02         A
#> sample03         A
#> sample04         A
#> sample05         A
#> ...            ...
#> sample08         B
#> sample09         B
#> sample10         B
#> sample11         B
#> sample12         B

## Create a copy of our minimal example.
y <- x
colnames(y) <- paste0("sample", seq(from = ncol(y) + 1L, to = ncol(y) * 2L))
colnames(y)

#>  [1] "sample13" "sample14" "sample15" "sample16" "sample17" "sample18"
#>  [7] "sample19" "sample20" "sample21" "sample22" "sample23" "sample24"

SummarizedExperiment::colData(y)

#> DataFrame with 12 rows and 1 column
#>          condition
#>           <factor>
#> sample13         A
#> sample14         A
#> sample15         A
#> sample16         A
#> sample17         A
#> ...            ...
#> sample20         B
#> sample21         B
#> sample22         B
#> sample23         B
#> sample24         B

## Combine two SummarizedExperiment objects.
c <- combine(x, y)
print(c)

#> class: RangedSummarizedExperiment 
#> dim: 500 24 
#> metadata(6): interestingGroups version ... wd sessionInfo
#> assays(1): counts
#> rownames(500): gene001 gene002 ... gene499 gene500
#> rowData names(5): geneID geneName geneBiotype broadClass entrezID
#> colnames(24): sample01 sample02 ... sample23 sample24
#> colData names(1): condition

colnames(c)

#>  [1] "sample01" "sample02" "sample03" "sample04" "sample05" "sample06"
#>  [7] "sample07" "sample08" "sample09" "sample10" "sample11" "sample12"
#> [13] "sample13" "sample14" "sample15" "sample16" "sample17" "sample18"
#> [19] "sample19" "sample20" "sample21" "sample22" "sample23" "sample24"

SummarizedExperiment::colData(c)

#> DataFrame with 24 rows and 1 column
#>          condition
#>           <factor>
#> sample01         A
#> sample02         A
#> sample03         A
#> sample04         A
#> sample05         A
#> ...            ...
#> sample20         B
#> sample21         B
#> sample22         B
#> sample23         B
#> sample24         B

## SingleCellExperiment ====
x <- sce
head(colnames(x))

#> [1] "cell0001" "cell0002" "cell0003" "cell0004" "cell0005" "cell0006"

sampleData(x)

#> DataFrame with 2 rows and 2 columns
#>         sampleName interestingGroups
#>           <factor>          <factor>
#> sample1    sample1           sample1
#> sample2    sample2           sample2

## Here we're faking a distinct replicate, just as an example.
y <- x
## Increase the cell ID numbers.
cells <- colnames(y) %>%
    sub("cell", "", .) %>%
    as.integer() %>%
    `+`(ncol(y)) %>%
    paste0("cell", .)
colnames(y) <- cells
head(colnames(y))

#> [1] "cell101" "cell102" "cell103" "cell104" "cell105" "cell106"

## Increase the sample ID numbers.
y$sampleID <- gsub("1$", "3", y$sampleID)
y$sampleID <- gsub("2$", "4", y$sampleID)
sampleData(y)

#> DataFrame with 2 rows and 2 columns
#>          sampleName interestingGroups
#>         <character>          <factor>
#> sample3     sample3           sample3
#> sample4     sample4           sample4

## Combine two SingleCellExperiment objects.
c <- combine(x, y)
print(c)

#> class: SingleCellExperiment 
#> dim: 500 200 
#> metadata(4): combine date wd sessionInfo
#> assays(1): counts
#> rownames(500): ENSG00000000003 ENSG00000000005 ... ENSG00000033050
#>   ENSG00000033100
#> rowData names(4): geneID geneName geneBiotype broadClass
#> colnames(200): cell0001 cell0002 ... cell199 cell200
#> colData names(2): expLibSize sampleID
#> reducedDimNames(0):
#> spikeNames(0):

sampleData(c)

#> DataFrame with 4 rows and 2 columns
#>         sampleName interestingGroups
#>           <factor>          <factor>
#> sample1    sample1           sample1
#> sample2    sample2           sample2
#> sample3    sample3           sample3
#> sample4    sample4           sample4