Convert Ensembl Identifiers to Gene Symbols

Convert Ensembl Identifiers to Gene Symbols

convertGenesToSymbols(object, ...)

# S4 method for character
convertGenesToSymbols(object, gene2symbol, organism,
  genomeBuild = NULL, release = NULL)

# S4 method for matrix
convertGenesToSymbols(object, ...)

# S4 method for data.frame
convertGenesToSymbols(object, ...)

# S4 method for DataFrame
convertGenesToSymbols(object, ...)

# S4 method for dgCMatrix
convertGenesToSymbols(object, ...)

# S4 method for dgTMatrix
convertGenesToSymbols(object, ...)

# S4 method for SummarizedExperiment
convertGenesToSymbols(object)

Arguments

object
Object.

...
Additional arguments.

gene2symbol
Optional. Gene-to-symbol mappings. If NULL, will attempt to download from Ensembl using the desired organism, genomeBuild, and release arguments.

organism
Optional. Organism name. Normally this argument is unnecessary and can be left unset. If a count matrix starts with a FASTA spike-in (e.g. "EGFP"), then automatic genome detection based on the first gene identifier will fail. In this case, the desired organism must be manually declared.

genomeBuild
Optional. Genome build assembly name (e.g. "GRCh38"). If NULL, defaults to the most recent build available.

release
Optional. Release version (e.g. 90). If NULL, defaults to the most recent release available.

Value

Same class as original object.

See also

Examples

# character ====
x <- c("ENSMUSG00000000001", "ENSMUSG00000000003")
convertGenesToSymbols(x)

#> Obtaining gene-to-symbol mappings from Ensembl

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 92 EnsDb for Mus Musculus

#> id: "AH60992"
#> organism: "Mus musculus"
#> build: "GRCm38"
#> release: "92"
#> format: "genes"

#> Standardizing the metadata columns

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

#> ENSMUSG00000000001 ENSMUSG00000000003 
#>            "Gnai3"             "Pbsn" 

# matrix ====
mat <- matrix(
    data = seq(1L:4L),
    byrow = TRUE,
    nrow = 2L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUSG00000000001", "ENSMUSG00000000003"),
        c("sample_1", "sample_2")
    )
)
print(mat)

#>                    sample_1 sample_2
#> ENSMUSG00000000001        1        2
#> ENSMUSG00000000003        3        4

mat <- convertGenesToSymbols(mat)

#> Obtaining gene-to-symbol mappings from Ensembl

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 92 EnsDb for Mus Musculus

#> id: "AH60992"
#> organism: "Mus musculus"
#> build: "GRCm38"
#> release: "92"
#> format: "genes"

#> Standardizing the metadata columns

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

print(mat)

#>       sample_1 sample_2
#> Gnai3        1        2
#> Pbsn         3        4

rownames(mat)

#> ENSMUSG00000000001 ENSMUSG00000000003 
#>            "Gnai3"             "Pbsn" 

# SummarizedExperiment ====
x <- convertGenesToSymbols(rse_bcb)
show(x)

#> class: RangedSummarizedExperiment 
#> dim: 500 6 
#> metadata(30): version level ... devtoolsSessionInfo subset
#> assays(1): counts
#> rownames(500): Cox5a Comt ... AC157654.1 AC129085.1
#> rowData names(7): geneID geneName ... entrezID broadClass
#> colnames(6): control_rep1 control_rep2 ... fa_day7_rep2 fa_day7_rep3
#> colData names(25): sampleName day ... x5x3Bias xGC

head(rownames(x))

#> [1] "Cox5a"  "Comt"   "Dazap2" "Rpl13"  "Calm1"  "Ddt"