Export

Export data out of R and write to disk.

export(x, ...)

# S4 method for ANY
export(x, file, format, ...)

# S4 method for sparseMatrix
export(x, file, format)

# S4 method for SummarizedExperiment
export(x, name = NULL, dir = ".",
  compress = FALSE, humanize = FALSE, slotNames = c("assays",
  "colData", "rowData", "Gene2Symbol", "Ensembl2Entrez"))

# S4 method for SingleCellExperiment
export(x, name = NULL, dir = ".",
  compress = FALSE, humanize = FALSE, slotNames = c("assays",
  "colData", "rowData", "Gene2Symbol", "Ensembl2Entrez"))

Arguments

x
An object supporting dim to be written to disk.

...
Additional arguments.

file
character(1). File path. Specify file or format but not both.

format
character(1). An optional character string containing the file format, which can be used to override the format inferred from file, or in lieu of specifying file.

name
character(1). Name to use on disk. If left NULL, will use the name of the object instead.

dir
character(1). Directory path.

compress
logical(1). Apply gzip compression to all files.

humanize
logical(1). Automatically convert gene IDs to gene symbols in the rownames and sample IDs to sample names in the colnames.

slotNames
character. Names of slots to include when writing to disk.

Value

Invisible character. File path(s).

Details

This is a wrapper for rio::export that adds support for additional S4 classes in Bioconductor.

Row names

The standard rio::export call will drop rownames when exporting to CSV. The readr family of functions also never write rownames. This is a really poor default setting for handling genomic data, which often contain gene identifiers in the rownames. Here we're performing any internal tibble coercion step to ensure rownames are always moved to a "rowname" column in the CSV export.

See also

Examples

library(SummarizedExperiment)

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: S4Vectors

#> 
#> Attaching package: 'S4Vectors'

#> The following object is masked from 'package:base':
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: 'matrixStats'

#> The following objects are masked from 'package:Biobase':
#> 
#>     anyMissing, rowMedians

#> The following object is masked from 'package:basejump':
#> 
#>     rowRanges

#> Loading required package: BiocParallel

#> 
#> Attaching package: 'DelayedArray'

#> The following objects are masked from 'package:matrixStats':
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following objects are masked from 'package:base':
#> 
#>     aperm, apply

library(SingleCellExperiment)
data(rse, sce)

## matrix ====
rnaseq_counts <- SummarizedExperiment::assay(rse)
export(rnaseq_counts, format = "csv")

#> Exported rnaseq_counts.csv.

## sparseMatrix ====
single_cell_counts <- SummarizedExperiment::assay(sce)
export(single_cell_counts, format = "mtx")

#> Exported single_cell_counts.mtx, single_cell_counts.mtx.colnames, single_cell_counts.mtx.rownames.

## Clean up.
file.remove(c(
    "rnaseq_counts.csv",
    "single_cell_counts.mtx",
    "single_cell_counts.mtx.colnames",
    "single_cell_counts.mtx.rownames"
))

#> [1] TRUE TRUE TRUE TRUE

## SummarizedExperiment ====
export(rse, dir = "example")

#> Exporting assays: counts

#> Exported counts.csv.

#> Exported colData.csv.

#> Exported rowData.csv.

#> Exported Gene2Symbol.csv.

#> 4 genes don't map to Entrez.

#> 2 genes map to multiple Entrez IDs.

#> Returning with 1:1 mappings using oldest Entrez ID per gene.

#> Exported Ensembl2Entrez.csv.

#> Exported rse to /data01/git/basejump/docs/reference/example/rse.

## SingleCellExperiment ====
export(sce, dir = "example")

#> Exporting assays: counts

#> Exported counts.mtx, counts.mtx.colnames, counts.mtx.rownames.

#> Exported colData.csv.

#> Exported rowData.csv.

#> Exported Gene2Symbol.csv.

#> 4 genes don't map to Entrez.

#> 2 genes map to multiple Entrez IDs.

#> Returning with 1:1 mappings using oldest Entrez ID per gene.

#> Exported Ensembl2Entrez.csv.

#> Exported sce to /data01/git/basejump/docs/reference/example/sce.

#> $assays
#> $assays$counts
#>                                                                       matrix 
#>          "/data01/git/basejump/docs/reference/example/sce/assays/counts.mtx" 
#>                                                                     barcodes 
#> "/data01/git/basejump/docs/reference/example/sce/assays/counts.mtx.colnames" 
#>                                                                        genes 
#> "/data01/git/basejump/docs/reference/example/sce/assays/counts.mtx.rownames" 
#> 
#> 
#> $colData
#> [1] "/data01/git/basejump/docs/reference/example/sce/colData.csv"
#> 
#> $rowData
#> [1] "/data01/git/basejump/docs/reference/example/sce/rowData.csv"
#> 
#> $Gene2Symbol
#> [1] "/data01/git/basejump/docs/reference/example/sce/Gene2Symbol.csv"
#> 
#> $Ensembl2Entrez
#> [1] "/data01/git/basejump/docs/reference/example/sce/Ensembl2Entrez.csv"
#> 

#> Exporting reducedDims: