Geometric Mean

The geometric mean is the nth root of n products or e to the mean log of x. Useful for describing non-normal (i.e. geometric) distributions.

geometricMean(object, ...)

# S4 method for numeric
geometricMean(object, removeNA = TRUE,
  zeroPropagate = FALSE)

# S4 method for matrix
geometricMean(object)

# S4 method for data.frame
geometricMean(object)

Arguments

object
Object.

...
Additional arguments.

removeNA
Remove NA values from calculations.

zeroPropagate
Allow propagation of zeroes.

Value

numeric containing geometric means.

Details

This function should be fully zero- and NA-tolerant. This calculation is not particularly useful if there are elements that are <= 0 and will return NaN.

See also

Examples

# numeric ====
vec <- seq(1L, 5L, 1L)
geometricMean(vec)

#> [1] 2.605171

vec2 <- vec ^ 2L
geometricMean(vec2)

#> [1] 6.786916

# data.frame ====
df <- data.frame(vec, vec2)
geometricMean(df)

#>      vec     vec2 
#> 2.605171 6.786916 

# matrix ====
mat <- as.matrix(df)
geometricMean(mat)

#>      vec     vec2 
#> 2.605171 6.786916