Get EnsDb from Bioconductor AnnotationHub

Get EnsDb from Bioconductor AnnotationHub

getEnsDb(organism, genomeBuild = NULL, release = NULL)

Arguments

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

genomeBuild
character(1). Ensembl genome build assembly name (e.g. "GRCh38"). If set NULL, defaults to the most recent build available. Note: don't pass in UCSC build IDs (e.g. "hg38").

release
integer(1). Ensembl release version (e.g. 90). We recommend setting this value if possible, for improved reproducibility. When left unset, the latest release available via AnnotationHub/ensembldb is used. Note that the latest version available can vary, depending on the versions of AnnotationHub and ensembldb in use.

Value

EnsDb.

Details

Remaps UCSC genome build to Ensembl automatically, if necessary. Provides legacy support for GRCh37 (hg19).

Note

Updated 2020-09-25.

Examples

getEnsDb(organism = "Homo sapiens", release = 100L)

#> → Getting `EnsDb` from AnnotationHub 2.20.2 (2020-04-27).

#> ℹ AH79689: Ensembl 100 EnsDb for Homo sapiens.

#> EnsDb for Ensembl:
#> |Backend: SQLite
#> |Db type: EnsDb
#> |Type of Gene ID: Ensembl Gene ID
#> |Supporting package: ensembldb
#> |Db created by: ensembldb package from Bioconductor
#> |script_version: 0.3.5
#> |Creation time: Sun May  3 02:11:59 2020
#> |ensembl_version: 100
#> |ensembl_host: localhost
#> |Organism: Homo sapiens
#> |taxonomy_id: 9606
#> |genome_build: GRCh38
#> |DBSCHEMAVERSION: 2.1
#> | No. of genes: 68008.
#> | No. of transcripts: 250776.
#> |Protein data available.