Import

Read file by extension into R.

import(file, dataFrame = getOption("basejump.data.frame", "data.frame"),
  ...)

Arguments

file
character(1). File path.

dataFrame
character(1). Data frame class to return. Can be set globally using the basejump.data.frame option.

Current recommendations (by priority):

data.frame: Base R default. Generally recommended.

S3 class.

Allows rownames, but they're required and can't be set NULL.

See help(topic = "data.frame", package = "base") for details.

DataFrame: Recommended when working with Bioconductor packages.

S4 class; inherits DataTable virtual class.

Allows rownames, but they're optional and can be set NULL.

See help(topic = "DataFrame", package = "S4Vectors") for details.

tbl_df (tibble): Recommended when working with tidyverse packages.

S3 class; inherits data.frame.

Does not allow rownames.

See help(topic = "tibble", package = "tibble") for details.

data.table: Recommended when working with the data.table package.

S3 class; inherits data.frame.

Does not allow rownames.

See help(topic = "data.table", package = "data.table") for details.

...
Additional arguments.

Value

Varies, depending on the file extension:

Default: DataFrame.

MTX: sparseMatrix.

GTF/GFF: GRanges.

JSON/YAML: list.

Source code or log: character.

Details

Remote URLs and compressed files are supported. All extensions are case insensitive.

This function supports automatic loading of common file types:

CSV: Comma Separated Values.
Imported by data.table::fread.

TSV Tab Separated Values.
Imported by data.table::fread.

TXT: Text file. Ambiguous, not recommended.
Imported by data.table::fread.

XLSX/XLS: Excel workbook.
Imported by readxl::read_excel.

MTX: MatrixMarket sparse matrix.
Imported by Matrix::readMM.

GTF/GFF/GFF3: General Feature Format.
Imported by rtracklayer::import.

JSON: JSON. Imported by jsonlite::read_json.

YAML/YML: YAML. Imported by yaml::yaml.load_file

RDA/RDATA: R Data.

Imported by load.

Must contain a single object.

Doesn't require internal object name to match, unlike loadData.

RDS: R Data Serialized.
Imported by readRDS.

These file formats will be imported as source code lines by readr::read_lines: LOG, MD, PY, R, RMD, SH.

These file formats are blacklisted, and intentionally not supported: DOC, DOCX, PDF, PPT, PPTX.

If a file format isn't supported natively (or blacklisted), the rio package will be used as a fallback attempt. See rio::import for details.

Delimited Files (CSV/TSV)

import uses the fread function of the data.table package to import standard CSV and TSV files. This should work automatically for most files without issue.

Here are some notable exceptions:

Columns headers should be character. In the event that a column name is numeric, set header = TRUE here to force the column name.

See help(topic = "fread", package = "data.table") for details.

The read_csv and read_tsv functions of the readr package are good alternatives, which return tibble data frames (tbl_df).

Matrix Market Exchange (MTX/MEX)

Reading a Matrix Market Exchange (MTX) file now requires COLNAMES and ROWNAMES sidecar files containing the colnames and rownames of the sparse matrix. Legacy support for manual loading of these sidecar files is provided.

General Feature Format (GFF/GTF)

The GFF (General Feature Format) format consists of one line per feature, each containing 9 columns of data, plus optional track definition lines. The GTF (General Transfer Format) is identical to GFF version 2.

Column names follow the Ensembl conventions. Additional information:

Ensembl

Gencode

bcbio files

Also supports some additional extensions commonly used with the bcbio pipeline:

COUNTS: Counts table (e.g. RNA-seq aligned counts).

COLNAMES: Sidecar file containing column names.

ROWNAMES: Sidecar file containing row names.

See also

Examples

url <- basejumpCacheURL

## Comma Separated Values
x <- import(file = file.path(url, "example.csv"))

#> Importing example.csv using data.table::fread().

#> Setting rownames from `rowname` column.

print(x)

#>         genotype treatment
#> sample1 wildtype   control
#> sample2 knockout   control
#> sample3 wildtype   treated
#> sample4 knockout   treated

## Tab Separated Values
x <- import(file = file.path(url, "example.tsv"))

#> Importing example.tsv using data.table::fread().

#> Setting rownames from `rowname` column.

## Microsoft Excel Worksheet
x <- import(file = file.path(url, "example.xlsx"))

#> Importing example.xlsx using readxl::read_excel().

#> Setting rownames from `rowname` column.

print(x)

#>         genotype treatment
#> sample1 wildtype   control
#> sample2 knockout   control
#> sample3 wildtype   treated
#> sample4 knockout   treated

## R Data
x <- import(file = file.path(url, "rnaseq_counts.rda"))

#> Importing rnaseq_counts.rda using base::load().

class(x)

#> [1] "matrix"

## GTF/GFF
# x <- import(file = file.path(url, "example.gtf"))
# summary(x)

## JSON
# x <- import(file = file.path(url, "example.json"))
# names(x)

## YAML
# x <- import(file = file.path(url, "example.yml"))
# names(x)

## Counts Table (i.e. aligned counts from bcbio)
# x <- import(file = file.path(url, "example.counts"))
# colSums(x)