Load Local Data

Load R data (.rda) files from a directory using symbols rather than complete file paths.

loadData(..., dir = getwd(), envir = parent.frame(), quiet = FALSE)

Arguments

...
Object names as symbols.

dir
Output directory. Defaults to the current working directory.

envir
Environment to use for assignment. Defaults to parent.frame(), which will assign into the calling environment.

quiet
If TRUE, suppress any status messages and/or progress bars.

Value

Silent named character vector of file paths.

Details

loadData() is opinionated about the format of R data files it will accept. base::save() allows for the saving of multiple objects into a single R data file. This can later result in unexpected accidental replacement of an existing object in the current environment. Additionally, since an R data file internally stores the name of an object, if the file is later renamed the object name will no longer match.

To avoid any accidental replacements, loadData() will only load R data files that contain a single object, and the internal object name must match the file name exactly. These conventions match the recommendations of the RStudio team, which recommends saving single objects per file.

Examples

# Load the internal GRCh37 gene annotations data
loadData(grch37, dir = system.file("extdata", package = "basejump"))

#> Loading grch37.rda from /Users/mike/GitHub/steinbaugh/basejump/inst/extdata

glimpse(grch37)

#> Observations: 63,677
#> Variables: 10
#> $ ensgene     <chr> "ENSG00000000003", "ENSG00000000005", "ENSG00000000419"...
#> $ symbol      <chr> "TSPAN6", "TNMD", "DPM1", "SCYL3", "C1orf112", "FGR", "...
#> $ description <chr> "tetraspanin 6 [Source:HGNC Symbol;Acc:11858]", "tenomo...
#> $ biotype     <chr> "protein_coding", "protein_coding", "protein_coding", "...
#> $ broadClass  <chr> "coding", "coding", "coding", "coding", "coding", "codi...
#> $ chr         <chr> "X", "X", "20", "1", "1", "1", "1", "6", "6", "6", "1",...
#> $ start       <int> 99883667, 99839799, 49551404, 169818772, 169631245, 279...
#> $ end         <int> 99894988, 99854882, 49575092, 169863408, 169823221, 279...
#> $ strand      <int> -1, 1, -1, -1, 1, -1, 1, -1, -1, 1, -1, 1, -1, 1, 1, 1,...
#> $ entrez      <list> [7105, 64102, 8813, 57147, 55732, 2268, 3075, 2519, 27...