Dynamically Handle a Local or Remote File Path

Dynamically Handle a Local or Remote File Path

localOrRemoteFile(object, severity = "stop", quiet = FALSE)

Arguments

object
Object.

severity
Return stop (default), warning, or message if file doesn't exist.

quiet
If TRUE, suppress any status messages and/or progress bars.

Value

Named character vector containing the original file name as the name and local file path as the string. Aborts on a missing file.

Examples

localOrRemoteFile("http://basejump.seq.cloud/mtcars.rda")

#>                                                                             mtcars.rda 
#> "/private/var/folders/j1/hw6j14mj3wq381x7mynwfbp00000gn/T/Rtmp7uROhN/file16ec41595564"