Gene-to-Symbol Mappings from Ensembl

Gene-to-Symbol Mappings from Ensembl

makeGene2symbolFromEnsembl(organism, genomeBuild = NULL, release = NULL)

Arguments

organism
Default recommended usage is to provide the full latin organism name as a string.

genomeBuild
Optional. Genome assembly name (e.g. GRCh38).

release
Optional. Ensembl release version. Defaults to the most current release available on AnnotationHub.

Value

data.frame