Map Genes

Take a user-defined gene vector and dynamically map the input to either the object rownames or the gene names (symbols). These functions are useful for writing code that needs to handle either gene identifier or gene name input dynamically (e.g. for single-cell RNA-seq marker analysis).

mapGenesToIDs(object, ...)

mapGenesToRownames(object, ...)

mapGenesToSymbols(object, ...)

# S4 method for SummarizedExperiment
mapGenesToRownames(object, genes)

# S4 method for SummarizedExperiment
mapGenesToIDs(object, genes)

# S4 method for SummarizedExperiment
mapGenesToSymbols(object, genes)

Arguments

object
Object.

...
Additional arguments.

Value

character.

Ambiguous gene names

Some genomes (e.g. Homo sapiens, Mus musculus) contain duplicated gene names for multiple gene identifiers. Normally we handle these ambiguous gene names by sanitizing them with base::make.names(). If a user requests a gene name that is duplicated, these functions will return a warning.

See also

Examples

object <- rse_small
head(rownames(object))

#> [1] "ENSG00000000003" "ENSG00000000005" "ENSG00000000419"
#> [4] "ENSG00000000457" "ENSG00000000460" "ENSG00000000938"

object_symbols <- convertGenesToSymbols(object)
head(rownames(object_symbols))

#> [1] "TSPAN6"   "TNMD"     "DPM1"     "SCYL3"    "C1orf112" "FGR"     

geneIDs <- head(rownames(object))
print(geneIDs)

#> [1] "ENSG00000000003" "ENSG00000000005" "ENSG00000000419"
#> [4] "ENSG00000000457" "ENSG00000000460" "ENSG00000000938"

geneNames <- head(as.character(rowRanges(object)$geneName))
print(geneNames)

#> [1] "TSPAN6"   "TNMD"     "DPM1"     "SCYL3"    "C1orf112" "FGR"     

# Row names
mapGenesToRownames(object, genes = geneIDs)

#> [1] "ENSG00000000003" "ENSG00000000005" "ENSG00000000419"
#> [4] "ENSG00000000457" "ENSG00000000460" "ENSG00000000938"

mapGenesToRownames(object, genes = geneNames)

#>            TSPAN6              TNMD              DPM1 
#> "ENSG00000000003" "ENSG00000000005" "ENSG00000000419" 
#>             SCYL3          C1orf112               FGR 
#> "ENSG00000000457" "ENSG00000000460" "ENSG00000000938" 

mapGenesToRownames(object_symbols, genes = geneIDs)

#> ENSG00000000003 ENSG00000000005 ENSG00000000419 ENSG00000000457 
#>        "TSPAN6"          "TNMD"          "DPM1"         "SCYL3" 
#> ENSG00000000460 ENSG00000000938 
#>      "C1orf112"           "FGR" 

mapGenesToRownames(object_symbols, genes = geneNames)

#> [1] "TSPAN6"   "TNMD"     "DPM1"     "SCYL3"    "C1orf112" "FGR"     

# Gene identifiers
mapGenesToIDs(object, genes = geneIDs)

#> [1] "ENSG00000000003" "ENSG00000000005" "ENSG00000000419"
#> [4] "ENSG00000000457" "ENSG00000000460" "ENSG00000000938"

mapGenesToIDs(object, genes = geneNames)

#>            TSPAN6              TNMD              DPM1 
#> "ENSG00000000003" "ENSG00000000005" "ENSG00000000419" 
#>             SCYL3          C1orf112               FGR 
#> "ENSG00000000457" "ENSG00000000460" "ENSG00000000938" 

mapGenesToIDs(object_symbols, genes = geneIDs)

#> [1] "ENSG00000000003" "ENSG00000000005" "ENSG00000000419"
#> [4] "ENSG00000000457" "ENSG00000000460" "ENSG00000000938"

mapGenesToIDs(object_symbols, genes = geneNames)

#>            TSPAN6              TNMD              DPM1 
#> "ENSG00000000003" "ENSG00000000005" "ENSG00000000419" 
#>             SCYL3          C1orf112               FGR 
#> "ENSG00000000457" "ENSG00000000460" "ENSG00000000938" 

# Gene names (symbols)
mapGenesToSymbols(object, genes = geneIDs)

#> ENSG00000000003 ENSG00000000005 ENSG00000000419 ENSG00000000457 
#>        "TSPAN6"          "TNMD"          "DPM1"         "SCYL3" 
#> ENSG00000000460 ENSG00000000938 
#>      "C1orf112"           "FGR" 

mapGenesToSymbols(object, genes = geneNames)

#> [1] "TSPAN6"   "TNMD"     "DPM1"     "SCYL3"    "C1orf112" "FGR"     

mapGenesToSymbols(object_symbols, genes = geneIDs)

#> ENSG00000000003 ENSG00000000005 ENSG00000000419 ENSG00000000457 
#>        "TSPAN6"          "TNMD"          "DPM1"         "SCYL3" 
#> ENSG00000000460 ENSG00000000938 
#>      "C1orf112"           "FGR" 

mapGenesToSymbols(object_symbols, genes = geneNames)

#> [1] "TSPAN6"   "TNMD"     "DPM1"     "SCYL3"    "C1orf112" "FGR"