Multiple Markdown tables

Knit multiple tables in a single R Markdown chunk.

markdownTables(list, captions = NULL, force = FALSE)

Arguments

list
Named list. Column data that can be coerced to data.frame.

captions
character or NULL. Table captions. If NULL, the names of the list will be used automatically as captions.

force
logical(1). Force knit output using knitr::asis_output(). Recommended for development and unit testing only.

Value

knit call: asis_output

Otherwise: list.

Note

knitr::kable() now supports multiple tables as a list for the x argument, but it still only supports a single caption. markdownTables extends this functionality, but supporting captions for each table.

See also

Examples

list <- list(
    mpg = head(ggplot2::mpg),
    mtcars = head(datasets::mtcars)
)
captions <- c(
    mpg = "Miles per gallon",
    mtcars = "Motor Trend car road tests"
)
markdownTables(list = list, captions = captions)

#> $mpg
#> # A tibble: 6 x 11
#>   manufacturer model displ  year   cyl trans      drv     cty   hwy fl    class 
#>   <chr>        <chr> <dbl> <int> <int> <chr>      <chr> <int> <int> <chr> <chr> 
#> 1 audi         a4      1.8  1999     4 auto(l5)   f        18    29 p     compa…
#> 2 audi         a4      1.8  1999     4 manual(m5) f        21    29 p     compa…
#> 3 audi         a4      2    2008     4 manual(m6) f        20    31 p     compa…
#> 4 audi         a4      2    2008     4 auto(av)   f        21    30 p     compa…
#> 5 audi         a4      2.8  1999     6 auto(l5)   f        16    26 p     compa…
#> 6 audi         a4      2.8  1999     6 manual(m5) f        18    26 p     compa…
#> 
#> $mtcars
#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1
#>