Match human gene orthologs

Match human gene orthologs

matchHumanOrthologs(genes, organism = NULL, ensemblRelease = NULL)

Arguments

genes
character. Gene identifiers.

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

ensemblRelease
integer(1). Ensembl release version (e.g. 90). We recommend setting this value if possible, for improved reproducibility. When left unset, the latest release available via AnnotationHub/ensembldb is used. Note that the latest version available can vary, depending on the versions of AnnotationHub and ensembldb in use.

Value

DataFrame. Data frame containing mapping columns:

geneID

hgncID

geneName

hgncName

Note

Updated 2019-08-18.

Examples

genes <- c(
    "ENSMUSG00000000001", "ENSMUSG00000000003",
    "ENSMUSG00000000028", "ENSMUSG00000000031",
    "ENSMUSG00000000037", "ENSMUSG00000000049"
)
matchHumanOrthologs(genes = genes, ensemblRelease = 87L)

#> Matching orthologs against Ensembl Genes 87 (dec2016.archive.ensembl.org) with biomaRt 2.40.3.

#> Getting Mus musculus gene symbols.

#> Making GRanges from Ensembl.

#> Matching EnsDb from AnnotationHub 2.16.0 (2019-05-02).

#> AH73905: Ensembl 97 EnsDb for Mus musculus
#> Run this code to download EnsDb manually:
#>   > library(AnnotationHub)
#>   > ah <- AnnotationHub()
#>   > edb <- ah[["AH73905"]]

#> Making GRanges from EnsDb object.

#>   - Organism: "Mus musculus"
#>   - Genome Build: "GRCm38"
#>   - Ensembl Release: 97L
#>   - Level: "genes"

#> Defining broadClass using: geneBiotype, geneName, seqnames.

#> Arranging by 'geneID'.

#> 56393 genes detected.

#> 768 non-unique gene symbols detected.

#> Returning with unmodified gene symbols (may contain duplicates).

#> Getting Homo sapiens gene symbols.

#> Making GRanges from Ensembl.

#> Matching EnsDb from AnnotationHub 2.16.0 (2019-05-02).

#> AH73986: Ensembl 79 EnsDb for Homo sapiens
#> Run this code to download EnsDb manually:
#>   > library(AnnotationHub)
#>   > ah <- AnnotationHub()
#>   > edb <- ah[["AH73986"]]

#> Making GRanges from EnsDb object.

#>   - Organism: "Homo sapiens"
#>   - Genome Build: "GRCh38"
#>   - Ensembl Release: 79L
#>   - Level: "genes"

#> Defining broadClass using: geneBiotype, geneName, seqnames.

#> Arranging by 'geneID'.

#> 65774 genes detected.

#> 2403 non-unique gene symbols detected.

#> Returning with unmodified gene symbols (may contain duplicates).

#> DataFrame with 6 rows and 4 columns
#>                                geneID    geneName          hgncID    hgncName
#>                           <character> <character>     <character> <character>
#> ENSMUSG00000000001 ENSMUSG00000000001       Gnai3 ENSG00000065135       GNAI3
#> ENSMUSG00000000003 ENSMUSG00000000003        Pbsn              NA          NA
#> ENSMUSG00000000028 ENSMUSG00000000028       Cdc45 ENSG00000093009       CDC45
#> ENSMUSG00000000031 ENSMUSG00000000031         H19              NA          NA
#> ENSMUSG00000000037 ENSMUSG00000000037       Scml2 ENSG00000102098       SCML2
#> ENSMUSG00000000049 ENSMUSG00000000049        Apoh ENSG00000091583        APOH