Check that interesting groups match a defined value

Prevent unwanted downstream behavior when a missing interesting group is requested by the user.

matchesInterestingGroups(x, interestingGroups, .xname = getNameInParent(x))

Arguments

x
Object.

interestingGroups
character. Groups of interest to use for visualization. Corresponds to factors describing the columns of the object.

.xname
Name of object defined in x. Not intended to be used directly.

Value

TRUE on success; FALSE on failure, with cause set.

Note

Updated 2019-08-11.

Examples

data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment

## Checks that columns are defined in `sampleData()`.
matchesInterestingGroups(rse, "condition")

#> [1] TRUE

## Currently allowing `NULL` to pass.
matchesInterestingGroups(rse, NULL)

#> [1] TRUE