Quality control metrics

Quality control metrics

metrics(object, ...)

metricsPerSample(object, ...)

# S4 method for SummarizedExperiment
metrics(object, return = c("tbl_df",
  "DataFrame"))

# S4 method for SingleCellExperiment
metrics(object, return = c("tbl_df",
  "DataFrame"))

# S4 method for SingleCellExperiment
metricsPerSample(object, fun = c("mean",
  "median", "sum"), return = c("tbl_df", "DataFrame"))

Arguments

object
Object.

return
character(1). Return type. Uses match.arg() internally and defaults to the first argument in the character vector.

fun
character(1). Mathematical function name to apply. Uses match.arg() internally.

...
Additional arguments.

Value

Object of class determined by return argument.

Details

metrics() takes data stored in colData() and consistently returns a tbl_df or DataFrame with sampleName and interestingGroups columns, even when these columns are not defined in colData(). This is designed to integrate with plotting functions that use ggplot2 internally.

Methods (by class)

SummarizedExperiment: Sample-level metrics.

SingleCellExperiment: Cell-level metrics.

SingleCellExperiment: Sample-level metrics.

Note

These functions will error intentionally if no numeric columns are defined in colData().

Updated 2019-08-18.

Examples

data(
    RangedSummarizedExperiment,
    SingleCellExperiment,
    package = "acidtest"
)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
object <- calculateMetrics(object)

#> Calculating 12 sample metrics.

#> 497 coding features detected.

#> 0 mitochondrial features detected.

x <- metrics(object)
print(x)

#> # A tibble: 12 x 10
#>    sampleID condition nCount nFeature nCoding nMito log10FeaturesPe… mitoRatio
#>    <chr>    <fct>      <int>    <int>   <int> <int>            <dbl>     <dbl>
#>  1 sample01 A          19139      445   19081    NA            0.618        NA
#>  2 sample02 A          19511      437   19473    NA            0.615        NA
#>  3 sample03 A          19047      449   19009    NA            0.620        NA
#>  4 sample04 A          18955      455   18891    NA            0.621        NA
#>  5 sample05 A          19449      452   19416    NA            0.619        NA
#>  6 sample06 A          19445      446   19423    NA            0.618        NA
#>  7 sample07 B          29796      459   29782    NA            0.595        NA
#>  8 sample08 B          30051      456   30037    NA            0.594        NA
#>  9 sample09 B          28145      444   28131    NA            0.595        NA
#> 10 sample10 B          27790      443   27776    NA            0.596        NA
#> 11 sample11 B          29351      445   29337    NA            0.593        NA
#> 12 sample12 B          24991      442   24974    NA            0.602        NA
#> # … with 2 more variables: sampleName <fct>, interestingGroups <fct>

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- calculateMetrics(object)

#> Calculating 100 sample metrics.

#> 497 coding features detected.

#> 0 mitochondrial features detected.

x <- metrics(object)
print(x)

#> # A tibble: 100 x 10
#>    cellID sampleID nCount nFeature nCoding nMito log10FeaturesPe… mitoRatio
#>    <chr>  <fct>     <int>    <int>   <int> <int>            <dbl>     <dbl>
#>  1 cell0… sample1   41054      235   40988    NA            0.514        NA
#>  2 cell0… sample2   49758      246   49642    NA            0.509        NA
#>  3 cell0… sample1   57439      248   57321    NA            0.503        NA
#>  4 cell0… sample2   68140      263   67976    NA            0.501        NA
#>  5 cell0… sample2   63809      245   63684    NA            0.497        NA
#>  6 cell0… sample1   49845      231   49701    NA            0.503        NA
#>  7 cell0… sample1   58662      245   58662    NA            0.501        NA
#>  8 cell0… sample1   43157      226   43049    NA            0.508        NA
#>  9 cell0… sample1   47298      229   47204    NA            0.505        NA
#> 10 cell0… sample1   44743      231   44743    NA            0.508        NA
#> # … with 90 more rows, and 2 more variables: sampleName <fct>,
#> #   interestingGroups <fct>

x <- metricsPerSample(object, fun = "mean")

#> Calculating mean per sample.

print(x)

#> # A tibble: 2 x 9
#>   sampleID sampleName interestingGrou… nCount nFeature nCoding nMito
#>   <chr>    <fct>      <fct>             <dbl>    <dbl>   <dbl> <dbl>
#> 1 sample1  sample1    sample1          55883.     240   55784.    NA
#> 2 sample2  sample2    sample2          53993.     245.  53886.    NA
#> # … with 2 more variables: log10FeaturesPerCount <dbl>, mitoRatio <dbl>