Generate syntactically valid names

These are convenience function that set names() on an object and return the object, without modifying the underlying data.

titleCase(character)

title_case(character)

firstCase(character)

first_case(character)

makeNames(character)

make_names(character)

makeNamesCamel(character)

make_names_camel(character)

camel(character)

makeNamesSnake(character)

make_names_snake(character)

snake(character)

setNamesCamel(object)

set_names_camel(object)

setNamesDot(object)

set_names_dot(object)

setNamesSnake(object)

set_names_snake(object)

sanitizeNames(object)

sanitize_names(object)

sanitiseNames(object)

sanitise_names(object)

Arguments

character
Character vector.

object
Object for which a names() attribute will be meaningful.

Value

Character vector or object with valid names.

Details

The makeNames() and *Case() family of functions manipulate a character vector, supporting either camelCase or snake_case.

The setNames() family of functions work similarly to makeNames() overall, but are to be used on objects supporting name assignments (e.g. data.frame, matrix, list).

sanitizeNames() modifies both colnames() and rownames() to snake_case format, and is generally recommended for clean-up of RNA-seq count matrices containing gene names.

See also

magrittr::set_names().

stats::setNames().

Examples

# Create character vectors of valid names
camel("RNAi clone")

#> [1] "rnaiClone"

snake("RNAi clone")

#> [1] "rnai_clone"

titleCase("RNA sequencing")

#> [1] "RNA Sequencing"

firstCase("RNA sequencing")

#> [1] "Rna Sequencing"

# Set names on objects
setNamesCamel(head(starwars))

#> # A tibble: 6 x 13
#>             name height  mass   hairColor   skinColor eyeColor birthYear gender
#>            <chr>  <int> <dbl>       <chr>       <chr>    <chr>     <dbl>  <chr>
#> 1 Luke Skywalker    172    77       blond        fair     blue      19.0   male
#> 2          C-3PO    167    75        <NA>        gold   yellow     112.0   <NA>
#> 3          R2-D2     96    32        <NA> white, blue      red      33.0   <NA>
#> 4    Darth Vader    202   136        none       white   yellow      41.9   male
#> 5    Leia Organa    150    49       brown       light    brown      19.0 female
#> 6      Owen Lars    178   120 brown, grey       light     blue      52.0   male
#> # ... with 5 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>

setNamesSnake(head(starwars))

#> # A tibble: 6 x 13
#>             name height  mass  hair_color  skin_color eye_color birth_year
#>            <chr>  <int> <dbl>       <chr>       <chr>     <chr>      <dbl>
#> 1 Luke Skywalker    172    77       blond        fair      blue       19.0
#> 2          C-3PO    167    75        <NA>        gold    yellow      112.0
#> 3          R2-D2     96    32        <NA> white, blue       red       33.0
#> 4    Darth Vader    202   136        none       white    yellow       41.9
#> 5    Leia Organa    150    49       brown       light     brown       19.0
#> 6      Owen Lars    178   120 brown, grey       light      blue       52.0
#> # ... with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>

setNamesDot(head(starwars))

#> # A tibble: 6 x 13
#>             name height  mass  hair.color  skin.color eye.color birth.year
#>            <chr>  <int> <dbl>       <chr>       <chr>     <chr>      <dbl>
#> 1 Luke Skywalker    172    77       blond        fair      blue       19.0
#> 2          C-3PO    167    75        <NA>        gold    yellow      112.0
#> 3          R2-D2     96    32        <NA> white, blue       red       33.0
#> 4    Darth Vader    202   136        none       white    yellow       41.9
#> 5    Leia Organa    150    49       brown       light     brown       19.0
#> 6      Owen Lars    178   120 brown, grey       light      blue       52.0
#> # ... with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>

# Sanitize both colnames and rownames
sanitizeNames(head(mtcars))

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> mazda_rx4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> mazda_rx4_wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> datsun_710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> hornet_4_drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> hornet_sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1