Parameters

Parameters

Arguments

alpha
scalar numeric or NULL. Adjusted P value ("alpha") cutoff. If left NULL, will use the cutoff defined in the object.

aspectRatio
scalar integer. Aspect ratio.

assay
scalar. Name or index of count matrix slotted in SummarizedExperiment::assays(). When passing in a string, the name must be defined in SummarizedExperiment::assayNames().

bpparam
bpparamClass. BiocParallel parameter specifying the back-end to be used for computations. See BiocParallel::bpparam() for details.

We recommend one of the following:

BiocParallel::bpparam().

BiocParallel::SerialParam().

BiocParallel::MulticoreParam().

check
boolean. Perform assert checks.

color
ggproto/ScaleDiscrete or NULL. Desired ggplot2 color scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_color_manual(). To set the discrete color palette globally, use options(basejump.color.discrete = ggplot2::scale_color_viridis_d()).

counts
matrix. Count matrix. Normalized counts are recommended.

countsAxisLabel
string. Counts axis label.

dark
boolean. Plot against a dark background using basejump::theme_midnight() ggplot2 theme.

dimsUse
integer. Vector of length 2 that denotes the columns from the reduced dimension matrix to use for centerX and centerY column calculations. Defaults the first and second dimensions.

dir
string. Directory path.

direction
string. Plot "both", "up", or "down" directions.

ensemblRelease
scalar integer. Ensembl release version (e.g. 90).

envir
environment to use for assignment. Defaults to parent.frame(), which will assign into the calling environment.

expression
string. Calculation to apply. Uses base::match.arg() and defaults to the first argument in the character vector.

file
string. File path.

fill
ggproto/ScaleDiscrete or NULL. Desired ggplot2 fill scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_fill_manual(). To set the discrete fill palette globally, use options(basejump.fill.discrete = scale_fill_viridis_d()).

flip
boolean. Flip x and y axes. Recommended for plots containing many samples.

gene2symbol
Gene2Symbol. Gene-to-symbol mappings. Must contain geneID and geneName columns. See Gene2Symbol() for more information.

genes
character. Gene identifiers. It is considered better practice to input the stable gene identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene symbols (e.g. "TSPN6"), if possible.

geom
string. Plot type. Uses base::match.arg() and defaults to the first argument in the character vector.

headerLevel
scalar integer (1-7). Markdown header level.

i
Indices specifying elements to extract or replace. Indices are numeric or character vectors, empty (missing), or NULL. See help(topic = "Extract", package = "base") for more information.

inherits
boolean. Should the enclosing frames of the environment be searched?

interestingGroups
character or NULL. Groups of interest that define the samples. If left unset, defaults to sampleName.

label
boolean. Superimpose sample text labels on the plot.

labelSize
scalar integer. Size of the text label.

lfcThreshold
scalar numeric or NULL. Log fold change ratio (base 2) cutoff threshold. If left NULL, will use the cutoff defined in the object.

legend
boolean. Show plot legend.

level
string. Return ranges as "genes" or "transcripts".

limit
scalar numeric. Threshold to denote on the plot, using a dashed line.

max
scalar numeric. Recommended maximum value cutoff.

min
scalar numeric. Recommended minimum value cutoff.

minCounts
scalar integer. Minimum number of counts per gene in the count matrix.

n
scalar integer. Number to include.

ntop
scalar integer. Number of top genes to label.

object
Object.

organism
string. Full Latin organism name (e.g. "Homo sapiens").

perMillion
boolean. Display as counts per million.

perSample
boolean. Visualize the distributions per sample.

plotlist
list containing ggplot objects.

pointAlpha
scalar numeric (0-1). Alpha transparency level. Useful when there many points in the dataset (e.g. single-cell data), and some points can be masked.

pointColor
string. Default point color for the plot.

pointScalar
scalar integer. Default point size for the plot.

pointsAsNumbers
boolean. Plot the points as numbers (TRUE) or dots (FALSE).

pointSize
scalar numeric. Point size for dots in the plot.

prefilter
boolean. Apply prefiltering to remove zero count genes.

progress
boolean. Show progress, using progress bars.

reducedDim
string. Name of reduced dimension matrix slotted in SingleCellExperiment::reducedDims(). Includes TNSE, UMAP, PCA, for example.

return
string. Return type. Uses base::match.arg() internally and defaults to the first argument in the character vector.

rowRanges
GRanges. Genomic ranges (e.g. genome annotations).

samples
character or NULL. Samples to include.

sigPointColor
character. Color names for labeling upregulated and downregulated genes. Also supports a character string for labeling DEGs with the same color, regardless of direction.

subtitle
string or NULL. Plot subtitle.

title
string or NULL. Plot title.

trans
string. Name of the axis scale transformation to apply. See help("scale_x_continuous", "ggplot2") for more information.

trendline
boolean. Include trendline on plot.

tx2gene
Tx2Gene. Transcript-to-gene mappings.

url
string. Uniform Resource Locator (URL). HTTP or FTP address.

value
Value to assign.

verbose
boolean. Run the function with verbose messages? It is only recommended to enable this when debugging.

x
Object.

...
Additional arguments.

Value

No value.