Parse GFF Attributes Column

Parse GFF Attributes Column

parseGFFAttributes(file, select = c("gene_", "transcript_"), unique = TRUE)

Arguments

file
File path.

select
Attribute to select. Supports partial matching. Defaults to returning attributes matching both "gene_" and "transcript_".

unique
Return unique attributes.

Value

tbl_df.

Note

Memory overhead and speed is greatly improved by requiring keys. Deparsing only unique values is much faster and generally recommended.

See also

Examples

parseGFFAttributes("http://basejump.seq.cloud/mmusculus.gtf") %>%
    glimpse()

#> Reading mmusculus.gtf

#> Observations: 20
#> Variables: 11
#> $ gene_biotype             <chr> "TEC", "snRNA", "protein_coding", "protein...
#> $ gene_id                  <chr> "ENSMUSG00000102693", "ENSMUSG00000064842"...
#> $ gene_name                <chr> "4933401J01Rik", "Gm26206", "Xkr4", "Xkr4"...
#> $ gene_source              <chr> "havana", "ensembl", "ensembl_havana", "en...
#> $ gene_version             <chr> "1", "1", "5", "5", "5", "1", "1", "1", "1...
#> $ transcript_biotype       <chr> "TEC", "snRNA", "processed_transcript", "p...
#> $ transcript_id            <chr> "ENSMUST00000193812", "ENSMUST00000082908"...
#> $ transcript_name          <chr> "4933401J01Rik-201", "Gm26206-201", "Xkr4-...
#> $ transcript_source        <chr> "havana", "ensembl", "havana", "havana", "...
#> $ transcript_support_level <chr> "NA", "NA", "1", "1", "1", "NA", "NA", "NA...
#> $ transcript_version       <chr> "1", "1", "1", "1", "4", "1", "1", "1", "1...