Prepare SummarizedExperiment

This is a utility wrapper for SummarizedExperiment() that provides automatic subsetting for colData and rowData.

prepareSummarizedExperiment(object, ...)

# S4 method for ANY
prepareSummarizedExperiment(object, rowData, colData,
  metadata = NULL)

# S4 method for data.frame
prepareSummarizedExperiment(object, rowData, colData,
  metadata = NULL)

# S4 method for dgCMatrix
prepareSummarizedExperiment(object, rowData, colData,
  metadata = NULL)

# S4 method for list
prepareSummarizedExperiment(object, rowData, colData,
  metadata = NULL)

# S4 method for matrix
prepareSummarizedExperiment(object, rowData, colData,
  metadata = NULL)

# S4 method for SimpleList
prepareSummarizedExperiment(object, rowData, colData,
  metadata = NULL)

Arguments

object
Object supporting dimensions (base::dim()), or a list containing valid objects. For NGS experiments, a counts matrix is recommended, and can be passed in either dense (matrix) or sparse (dgCMatrix, dgTMatrix) format. Multiple matrices can be supplied as a list, as long as they all have the same dimensions. List object can be supplied as either class list or SimpleList.

...
Additional arguments (for the S4 generic definition).

rowData
Object describing assay matrix rows. Must support base::dim().

colData
Object describing assay matrix columns. Must support base::dim().

metadata
Optional. Metadata list.

Value

SummarizedExperiment.

Details

prepareSummarizedExperiment() also provides automatic metadata slotting of multiple useful environment parameters:

date: Today's date.

wd: Working directory.

sessionInfo: R session information.

See also

SummarizedExperiment::SummarizedExperiment().

base::Sys.Date().

base::getwd().

utils::sessionInfo().

Other Data Import and Project Utilities: loadRemoteData, readFileByExtension, readGFF, readYAML, transmit

Examples

mat <- matrix(
    seq(1L:16L),
    nrow = 4L,
    ncol = 4L,
    dimnames = list(
        c("gene_1", "gene_2", "gene_3", "gene_4"),
        c("sample_1", "sample_2", "sample_3", "sample_4")))
rowData <- data.frame(
    ensgene = c("Aaa", "Bbb", "Ccc", "Ddd"),
    biotype = c("coding", "coding", "coding", "pseudogene"),
    row.names = rownames(mat))
colData <- data.frame(
    genotype = c("wt", "wt", "ko", "ko"),
    age = c(3L, 6L, 3L, 6L),
    row.names = colnames(mat))
prepareSummarizedExperiment(
    mat,
    rowData = rowData,
    colData = colData)

#> Preparing SummarizedExperiment

#> class: SummarizedExperiment 
#> dim: 4 4 
#> metadata(3): date wd sessionInfo
#> assays(1): assay
#> rownames(4): gene_1 gene_2 gene_3 gene_4
#> rowData names(2): ensgene biotype
#> colnames(4): sample_1 sample_2 sample_3 sample_4
#> colData names(2): genotype age