Read File by Extension

Supports automatic loading of common file extensions:

readFileByExtension(file, makeNames = c("camel", "dotted", "snake",
  "upperCamel"), ...)

Arguments

file
File path.

makeNames
Make syntactically valid names. Supports camel, dotted, snake, or upperCamel. This is always enforced.

...
Additional arguments.

Value

tbl_df by default, or a sparse matrix for .mtx files.

Details

csv: Comma Separated Values.

gff/gff3/gtf: General Feature Format.

mtx: MatrixMarket sparse matrix.

rda/RData: R Data. Must contain a single object. Doesn't require internal object name to match, like in the loadData() function.

rds: R Data Serialized.

tsv Tab Separated Values.

xlsx: Excel workbook.

yaml/yml: YAML.

Also supports some additional extensions commonly used with the bcbio pipeline:

counts: Counts table.

colnames: Sidecar file containing column names.

rownames: Sidecar file containing row names.

If the file format isn't supported natively (or blacklisted), the rio package will be used as a fallback attempt.

Note

Reading a MatrixMarket ("mtx") file now requires "colnames" and "rownames" sidecar files containing the colnames() and rownames() of the sparse matrix. Legacy support for manual loading of these sidecar files is provided.

See also

Examples

# Comma Separated Values
x <- readFileByExtension("http://basejump.seq.cloud/mtcars.csv")

#> Reading mtcars.csv

#> Parsed with column specification:
#> cols(
#>   rowname = col_character(),
#>   mpg = col_double(),
#>   cyl = col_integer(),
#>   disp = col_double(),
#>   hp = col_integer(),
#>   drat = col_double(),
#>   wt = col_double(),
#>   qsec = col_double(),
#>   vs = col_integer(),
#>   am = col_integer(),
#>   gear = col_integer(),
#>   carb = col_integer()
#> )

glimpse(x)

#> Observations: 32
#> Variables: 12
#> $ rowname <chr> "Mazda RX4", "Mazda RX4 Wag", "Datsun 710", "Hornet 4 Drive...
#> $ mpg     <dbl> 21.0, 21.0, 22.8, 21.4, 18.7, 18.1, 14.3, 24.4, 22.8, 19.2,...
#> $ cyl     <int> 6, 6, 4, 6, 8, 6, 8, 4, 4, 6, 6, 8, 8, 8, 8, 8, 8, 4, 4, 4,...
#> $ disp    <dbl> 160.0, 160.0, 108.0, 258.0, 360.0, 225.0, 360.0, 146.7, 140...
#> $ hp      <int> 110, 110, 93, 110, 175, 105, 245, 62, 95, 123, 123, 180, 18...
#> $ drat    <dbl> 3.90, 3.90, 3.85, 3.08, 3.15, 2.76, 3.21, 3.69, 3.92, 3.92,...
#> $ wt      <dbl> 2.620, 2.875, 2.320, 3.215, 3.440, 3.460, 3.570, 3.190, 3.1...
#> $ qsec    <dbl> 16.46, 17.02, 18.61, 19.44, 17.02, 20.22, 15.84, 20.00, 22....
#> $ vs      <int> 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1,...
#> $ am      <int> 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1,...
#> $ gear    <int> 4, 4, 4, 3, 3, 3, 3, 4, 4, 4, 4, 3, 3, 3, 3, 3, 3, 4, 4, 4,...
#> $ carb    <int> 4, 4, 1, 1, 2, 1, 4, 2, 2, 4, 4, 3, 3, 3, 4, 4, 4, 1, 2, 1,...

# Counts Table (bcbio)
x <- readFileByExtension("http://basejump.seq.cloud/example.counts")

#> Reading example.counts

#> Parsed with column specification:
#> cols(
#>   id = col_character(),
#>   sample01 = col_integer(),
#>   sample02 = col_integer(),
#>   sample03 = col_integer(),
#>   sample04 = col_integer(),
#>   sample05 = col_integer(),
#>   sample06 = col_integer(),
#>   sample07 = col_integer(),
#>   sample08 = col_integer(),
#>   sample09 = col_integer(),
#>   sample10 = col_integer(),
#>   sample11 = col_integer(),
#>   sample12 = col_integer()
#> )

glimpse(x)

#>  int [1:9, 1:12] 0 0 259 0 3 0 0 0 0 0 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:9] "ENSMUSG00000102693" "ENSMUSG00000064842" "ENSMUSG00000051951" "ENSMUSG00000102851" ...
#>   ..$ : chr [1:12] "sample01" "sample02" "sample03" "sample04" ...

# R Data
x <- readFileByExtension("http://basejump.seq.cloud/rnaseqCounts.rda")

#> Reading rnaseqCounts.rda

glimpse(x)

#>  int [1:1000, 1:12] 17 0 21 3 10 1 0 26 17 7 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:1000] "gene1" "gene2" "gene3" "gene4" ...
#>   ..$ : chr [1:12] "sample1" "sample2" "sample3" "sample4" ...

# Microsoft Excel Worksheet
x <- readFileByExtension("http://basejump.seq.cloud/mtcars.xlsx")

#> Reading mtcars.xlsx

glimpse(x)

#> Observations: 6
#> Variables: 12
#> $ rowname <chr> "Mazda RX4", "Mazda RX4 Wag", "Datsun 710", "Hornet 4 Drive...
#> $ mpg     <dbl> 21.0, 21.0, 22.8, 21.4, 18.7, 18.1
#> $ cyl     <dbl> 6, 6, 4, 6, 8, 6
#> $ disp    <dbl> 160, 160, 108, 258, 360, 225
#> $ hp      <dbl> 110, 110, 93, 110, 175, 105
#> $ drat    <dbl> 3.90, 3.90, 3.85, 3.08, 3.15, 2.76
#> $ wt      <dbl> 2.620, 2.875, 2.320, 3.215, 3.440, 3.460
#> $ qsec    <dbl> 16.46, 17.02, 18.61, 19.44, 17.02, 20.22
#> $ vs      <dbl> 0, 0, 1, 1, 0, 1
#> $ am      <dbl> 1, 1, 1, 0, 0, 0
#> $ gear    <dbl> 4, 4, 4, 3, 3, 3
#> $ carb    <dbl> 4, 4, 1, 1, 2, 1

# Text Table (ambiguous; not recommended)
x <- suppressWarnings(
    readFileByExtension("http://basejump.seq.cloud/example.txt")
)

#> Reading example.txt

glimpse(x)

#> Observations: 32
#> Variables: 11
#> $ mpg  <dbl> 21.0, 21.0, 22.8, 21.4, 18.7, 18.1, 14.3, 24.4, 22.8, 19.2, 17...
#> $ cyl  <int> 6, 6, 4, 6, 8, 6, 8, 4, 4, 6, 6, 8, 8, 8, 8, 8, 8, 4, 4, 4, 4,...
#> $ disp <dbl> 160.0, 160.0, 108.0, 258.0, 360.0, 225.0, 360.0, 146.7, 140.8,...
#> $ hp   <int> 110, 110, 93, 110, 175, 105, 245, 62, 95, 123, 123, 180, 180, ...
#> $ drat <dbl> 3.90, 3.90, 3.85, 3.08, 3.15, 2.76, 3.21, 3.69, 3.92, 3.92, 3....
#> $ wt   <dbl> 2.620, 2.875, 2.320, 3.215, 3.440, 3.460, 3.570, 3.190, 3.150,...
#> $ qsec <dbl> 16.46, 17.02, 18.61, 19.44, 17.02, 20.22, 15.84, 20.00, 22.90,...
#> $ vs   <int> 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1,...
#> $ am   <int> 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0,...
#> $ gear <int> 4, 4, 4, 3, 3, 3, 3, 4, 4, 4, 4, 3, 3, 3, 3, 3, 3, 4, 4, 4, 3,...
#> $ carb <int> 4, 4, 1, 1, 2, 1, 4, 2, 2, 4, 4, 3, 3, 3, 4, 4, 4, 1, 2, 1, 1,...