Save Data

Wrapper for base::save() supporting quick saving of object names passed as symbols. This function saves each object into a separate .rda file rather than combining into a single file.

saveData(..., dir = ".", ext = getOption("basejump.save.ext", "rda"),
  overwrite = TRUE, compress = getOption("basejump.save.compress",
  TRUE))

Arguments

...
Object names. Note that these arguments are interpreted as symbols using non-standard evaluation for convenience during interactive use, and should not be quoted.

dir
string. Output directory. Defaults to the current working directory.

ext
string. R data ("rda", "RData") or R data serialized (RDS; "rds"). RDS is preferred when saving single objects per file, which is always the convention of saveData(), regardless of the extension used.

overwrite
boolean. Overwrite existing file.

compress
logical or character string specifying whether saving to a named file is to use compression. TRUE corresponds to gzip compression, and character strings "gzip", "bzip2" or "xz" specify the type of compression. Ignored when file is a connection and for workspace format version 1.

Value

Invisible named character. File paths.

Note

This function is desired for interactive use and interprets object names using non-standard evaluation. It will overwrite existing saved data, following the same conventions as base::save(). Conversely, devtools::use_data() does not overwrite by default if that behavior is preferred.

See also

Examples

saveData(rnaseq_counts, single_cell_counts, dir = "example")

#> Saving rnaseq_counts.rda, single_cell_counts.rda to /Users/mike/git/basejump-develop/docs/reference/example

list.files("example")

#> [1] "rnaseq_counts.rda"      "single_cell_counts.rda"

# Clean up
unlink("example", recursive = TRUE)