Snake Case

Snake Case

snake(object, ...)

# S4 method for ANY
snake(object, rownames = FALSE)

# S4 method for character
snake(object)

# S4 method for data.frame
snake(object, rownames = FALSE)

# S4 method for list
snake(object)

# S4 method for matrix
snake(object, rownames = FALSE)

# S4 method for tbl_df
snake(object)

Arguments

object
Character vector or an object for which names() assignment will be meaningful.

...
Additional arguments (for the S4 generic definition).

rownames
Apply sanitization on row names. This is not recommended by default, since rownames commonly contain gene identifiers that should not be modified.

Value

Object with syntatically valid names. For objects supporting base::names(), the underlying data returns unchanged.

Details

For unnamed character vectors, this function will sanitize the underlying values. Otherwise, it will set names() and/or rownames() on objects supporting name assignments. They return the object without modification of the underlying data.

Examples

loadRemoteData("http://basejump.seq.cloud/makeNames.rda")

# Character vector
snake(makeNames$vec)

#>  [1] "hello_world"       "hello_world"       "rnai_clones"      
#>  [4] "n_count"           "tx2gene"           "tx2_gene_id"      
#>  [7] "g2m_score"         "worfdb_html_remap" "mazda_rx4"        
#> [10] "x123"              NA                 

# Named character vector
snake(makeNames$namedVec)

#>        item_a        item_b 
#> "hello world" "HELLO WORLD" 

# data.frame
snake(makeNames$df, rownames = TRUE)

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> mazda_rx4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> mazda_rx4_wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> datsun_710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> hornet_4_drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> hornet_sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

# Named list
snake(makeNames$lst)

#> $item_a
#> [1] 1 2
#> 
#> $item_b
#> [1] 3 4
#>