Validate Classes

Validity check capable of validating multiple slots in a single call.

validateClasses(object, expected, subset = FALSE)

Arguments

object
Object.

expected
list. Named list of expected classes per slot.

subset
boolean. Only check a subset of slots in the object.

Value

boolean (TRUE) on sucess or string containing informative message on failure.

Details

To be used inside S4 methods::setValidity() call or with assertthat::validate_that(). Particularly useful for checking multiple slotted objects inside metadata().

See also

Examples

data(rse)
validateClasses(
    object = S4Vectors::metadata(rse),
    expected = list(
        version = c("package_version", "numeric_version"),
        date = "Date",
        interestingGroups = "character"
    )
)

#> [1] TRUE