Write Counts

Supports both bulk and single-cell RNA-seq output. Bulk RNA-seq counts are written to disk as comma separated values (.csv). Single-cell RNA-seq counts are written to disk in MatrixMarket format (.mtx), along with the sample barcodes (.colnames), and gene identifiers (.rownames).

writeCounts(..., dir = getwd(), gzip = TRUE, quiet = FALSE)

Arguments

...
Count matrices, passed in as dots.

dir
Output directory.

gzip
Compress the counts file using gzip.

quiet
If TRUE, suppress any status messages and/or progress bars.

Value

Invisible named character vector containing file paths to count matrices written to disk.

Details

Automatic gzip compression is offered as a user-defined option. This setting is enabled by default to save disk space. Note that the readr package, built into RStudio, now natively supports compressed files.

See also

Other Write Utilities: saveData

Examples

writeCounts(mtcars, gzip = TRUE)

#> Writing mtcars to /Users/mike/GitHub/steinbaugh/basejump/docs/reference

# Clean up
unlink("mtcars.csv.gz")