Ensembl Annotations
Source: R/annotable.R

Ensembl Annotations

annotable(organism, format = c("genes", "transcripts"), genomeBuild = NULL,
  release = NULL)

Arguments

organism
Full latin organism name (e.g. "Homo sapiens").

format
Fetch ranges as "genes" or "transcripts".

genomeBuild
Optional. Genome assembly name (e.g. "GRCh38"). If NULL, defaults to the most recent build available.

release
Optional. Release version (e.g. 90). If NULL, defaults to the most recent release available.

Value

data.frame.

Details

Simply specify the desired organism, using the full latin name. For example, we can obtain human annotations with Homo sapiens. Optionally, specific Ensembl genome builds (e.g. GRCh38) and release versions (e.g. 87) are supported.

Under the hood, this function fetches annotations from AnnotationHub using the ensembldb package. AnnotationHub supports versioned Ensembl releases, back to version 87.

Broad Class Definitions

For gene and transcript tables, when broadClass = TRUE, a broadClass column is added, which generalizes the gene types into a smaller number of semantically-meaningful groups:

coding

noncoding

pseudo

small

decaying

ig (immunoglobulin)

tcr (T cell receptor)

other

GRCh37 Legacy Annotations

The Homo sapiens GRCh37 (release 75) build is supported by internally querying the EnsDb.Hsapiens.v75 package.

See also

Examples

# Genes ====
x <- annotable("Homo sapiens")

#> Making GRanges from Ensembl with AnnotationHub 2.10.1 (2017-10-27)

#> id: "AH57757"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "90"
#> format: "genes"

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

glimpse(x)

#> Observations: 64,661
#> Variables: 12
#> $ seqnames       <fct> X, X, 20, 1, 1, 1, 1, 6, 6, 6, 1, 1, X, 6, 3, 7, 7, ...
#> $ start          <int> 100627109, 100584802, 50934867, 169849631, 169662007...
#> $ end            <int> 100639991, 100599885, 50958555, 169894267, 169854080...
#> $ width          <int> 12883, 15084, 23689, 44637, 192074, 23214, 95627, 16...
#> $ strand         <fct> -, +, -, -, +, -, +, -, -, +, -, +, -, +, +, +, +, -...
#> $ geneID         <chr> "ENSG00000000003", "ENSG00000000005", "ENSG000000004...
#> $ geneName       <chr> "TSPAN6", "TNMD", "DPM1", "SCYL3", "C1orf112", "FGR"...
#> $ geneBiotype    <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ description    <chr> "tetraspanin 6 [Source:HGNC Symbol;Acc:HGNC:11858]",...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...
#> $ entrezID       <I(list)>  7105, 64102,  8813, 57147, 55732,  2268,  3075,...
#> $ broadClass     <fct> coding, coding, coding, coding, coding, coding, codi...

# Transcripts ====
x <- annotable("Homo sapiens", format = "transcripts")

#> Making GRanges from Ensembl with AnnotationHub 2.10.1 (2017-10-27)

#> id: "AH57757"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "90"
#> format: "transcripts"

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Setting txBiotype as factor

#> Setting txSupportLevel as factor

#> Defining broadClass using: geneName, txBiotype, seqnames

glimpse(x)

#> Observations: 220,144
#> Variables: 18
#> $ seqnames       <fct> 7, 12, 11, 12, 2, 2, 6, 16, 4, 3, 7, 7, 11, 1, 1, 7,...
#> $ start          <int> 127588345, 8940365, 64305578, 2794953, 72129238, 372...
#> $ end            <int> 127591705, 8949955, 64316738, 2805423, 72148038, 372...
#> $ width          <int> 3361, 9591, 11161, 10471, 18801, 17495, 16880, 14609...
#> $ strand         <fct> +, -, +, +, -, +, -, -, -, +, +, -, -, -, +, +, +, -...
#> $ txID           <chr> "ENST00000000233", "ENST00000000412", "ENST000000004...
#> $ txName         <chr> "ENST00000000233", "ENST00000000412", "ENST000000004...
#> $ txBiotype      <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ geneID         <chr> "ENSG00000004059", "ENSG00000003056", "ENSG000001731...
#> $ geneName       <chr> "ARF5", "M6PR", "ESRRA", "FKBP4", "CYP26B1", "NDUFAF...
#> $ geneBiotype    <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ description    <chr> "ADP ribosylation factor 5 [Source:HGNC Symbol;Acc:H...
#> $ txCdsSeqStart  <int> 127588499, 8941818, 64307180, 2795140, 72132227, 372...
#> $ txCdsSeqEnd    <int> 127591299, 8946404, 64315966, 2803258, 72147834, 372...
#> $ txSupportLevel <fct> 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 5, 1, 1, 1, 1...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...
#> $ entrezID       <I(list)>    381,   4074,   2101,   2288,  56603,  55471, ...
#> $ broadClass     <fct> coding, coding, coding, coding, coding, coding, codi...