Ensembl Annotations

Quickly obtain gene and transcript annotations from Ensembl using AnnotationHub and ensembldb.

annotable(organism, format = c("genes", "transcripts"),
  genomeBuild = NULL, release = NULL)

Arguments

organism
string. Full latin organism name (e.g. "Homo sapiens").

format
string. Fetch ranges as "genes" or "transcripts".

genomeBuild
string or NULL. Genome build assembly name (e.g. "GRCh38"). If set NULL, defaults to the most recent build available.

release
scalar integer or NULL. Release version (e.g. 90). If set NULL, defaults to the most recent release available.

Value

data.frame.

Details

Simply specify the desired organism, using the full latin name. For example, we can obtain human annotations with Homo sapiens. Optionally, specific Ensembl genome builds (e.g. GRCh38) and release versions (e.g. 87) are supported.

Under the hood, this function fetches annotations from AnnotationHub using the ensembldb package. AnnotationHub supports versioned Ensembl releases, back to version 87.

Broad class definitions

For gene and transcript annotations, a broadClass column is added, which generalizes the gene types into a smaller number of semantically-meaningful groups:

coding

noncoding

pseudo

small

decaying

ig (immunoglobulin)

tcr (T cell receptor)

other

GRCh37 legacy annotations

The Homo sapiens GRCh37 (release 75) build is supported by internally querying the EnsDb.Hsapiens.v75 package.

See also

Examples

# Genes ====
x <- annotable("Homo sapiens")

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 92 EnsDb for Homo Sapiens

#> id: "AH60977"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "92"
#> format: "genes"

#> Standardizing the metadata columns

#> Defining broad class using: geneName, geneBiotype, seqnames

#> Sorting ranges by geneID

glimpse(x)

#> Observations: 65,256
#> Variables: 12
#> $ seqnames       <fct> X, X, 20, 1, 1, 1, 1, 6, 6, 6, 1, 1, X, 6, 3, 7, 7, ...
#> $ start          <int> 100627109, 100584802, 50934867, 169849631, 169662007...
#> $ end            <int> 100639991, 100599885, 50958555, 169894267, 169854080...
#> $ width          <int> 12883, 15084, 23689, 44637, 192074, 23214, 95627, 16...
#> $ strand         <fct> -, +, -, -, +, -, +, -, -, +, -, +, -, +, +, +, +, -...
#> $ broadClass     <fct> coding, coding, coding, coding, coding, coding, codi...
#> $ description    <fct> tetraspanin 6 [Source:HGNC Symbol;Acc:HGNC:11858], t...
#> $ entrezID       <I(list)>  7105, 64102,  8813, 57147, 55732,  2268,  3075,...
#> $ geneBiotype    <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ geneID         <chr> "ENSG00000000003", "ENSG00000000005", "ENSG000000004...
#> $ geneName       <fct> TSPAN6, TNMD, DPM1, SCYL3, C1orf112, FGR, CFH, FUCA2...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...

# Transcripts ====
x <- annotable("Homo sapiens", format = "transcripts")

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 92 EnsDb for Homo Sapiens

#> id: "AH60977"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "92"
#> format: "transcripts"

#> Standardizing the metadata columns

#> Defining broad class using: geneName, transcriptBiotype, seqnames

#> Sorting ranges by transcriptID

glimpse(x)

#> Observations: 225,358
#> Variables: 18
#> $ seqnames               <fct> 7, 12, 11, 12, 2, 2, 6, 16, 4, 3, 7, 7, 11, ...
#> $ start                  <int> 127588345, 8940365, 64305578, 2794953, 72129...
#> $ end                    <int> 127591705, 8949955, 64316738, 2805423, 72148...
#> $ width                  <int> 3361, 9591, 11161, 10471, 18801, 17495, 1688...
#> $ strand                 <fct> +, -, +, +, -, +, -, -, -, +, +, -, -, -, +,...
#> $ broadClass             <fct> coding, coding, coding, coding, coding, codi...
#> $ description            <fct> ADP ribosylation factor 5 [Source:HGNC Symbo...
#> $ entrezID               <I(list)>    381,   4074,   2101,   2288,  56603, ...
#> $ geneBiotype            <fct> protein_coding, protein_coding, protein_codi...
#> $ geneID                 <fct> ENSG00000004059, ENSG00000003056, ENSG000001...
#> $ geneName               <fct> ARF5, M6PR, ESRRA, FKBP4, CYP26B1, NDUFAF7, ...
#> $ seqCoordSystem         <fct> chromosome, chromosome, chromosome, chromoso...
#> $ transcriptBiotype      <fct> protein_coding, protein_coding, protein_codi...
#> $ transcriptCdsSeqEnd    <int> 127591299, 8946404, 64315966, 2803258, 72147...
#> $ transcriptCdsSeqStart  <int> 127588499, 8941818, 64307180, 2795140, 72132...
#> $ transcriptID           <chr> "ENST00000000233", "ENST00000000412", "ENST0...
#> $ transcriptName         <chr> "ENST00000000233", "ENST00000000412", "ENST0...
#> $ transcriptSupportLevel <int> 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 5, 1,...