Ensembl Annotations

Quickly obtain gene annotations from Ensembl.

annotable(object, ...)

# S4 method for character
annotable(object, format = "gene", release,
  quiet = FALSE)

# S4 method for data.frame
annotable(object)

Arguments

object
Object. Default recommended usage is to provide full latin organism name as a string. Also supports input of pre-built tibbles from the annotables data package.

...
Additional arguments (for the S4 generic definition).

format
Desired table format, either gene, tx2gene, or gene2symbol.

release
Optional. Ensembl release version. Defaults to the most current release available on AnnotationHub.

quiet
If TRUE, suppress any status messages and/or progress bars.

Value

data.frame with unique rows per gene or transcript.

Using AnnotationHub

The default recommended approach is to simply specify the desired organism, using the full latin name. For example, we can obtain human annotations with Homo sapiens and mouse annotations with Mus musculus. Under the hood, the function fetches these annotations from AnnotationHub using the ensembldb package. AnnotationHub supports versioned Ensembl releases, back to version 87.

Using Pre-Compiled Annotables

Generally we' don't recommend this approach. However, pass-in of a tibble generated using biomaRt in the annotables data package is useful for adding support for the outdated (but still frequently used) GRCh37/hg19 human genome.

Broad Class Definitions

A broadClass column is added, which generalizes the gene types into a smaller number of semantically-meaningful groups:

coding

noncoding

pseudo

small

decaying

ig (immunoglobulin)

tcr (T cell receptor)

other

See also

AnnotationHub.

ensembldb.

Other Gene Annotation Utilities: detectOrganism, gene2symbolFromGFF, gene2symbol, symbol2gene, tx2geneFromGFF, tx2gene

Examples

annotable("Mus musculus") %>% str()

#> Loading Ensembl annotations from AnnotationHub
#> 2017-10-27

#> EnsDB AH57770: Mus musculus Ensembl 90

#> 'data.frame':	53110 obs. of  10 variables:
#>  $ ensgene       : chr  "ENSMUSG00000000001" "ENSMUSG00000000003" "ENSMUSG00000000028" "ENSMUSG00000000031" ...
#>  $ symbol        : chr  "Gnai3" "Pbsn" "Cdc45" "H19" ...
#>  $ description   : chr  "guanine nucleotide binding protein (G protein), alpha inhibiting 3 [Source:MGI Symbol;Acc:MGI:95773]" "probasin [Source:MGI Symbol;Acc:MGI:1860484]" "cell division cycle 45 [Source:MGI Symbol;Acc:MGI:1338073]" "H19, imprinted maternally expressed transcript [Source:MGI Symbol;Acc:MGI:95891]" ...
#>  $ biotype       : chr  "protein_coding" "protein_coding" "protein_coding" "lincRNA" ...
#>  $ broadClass    : chr  "coding" "coding" "coding" "noncoding" ...
#>  $ geneSeqStart  : int  108107280 77837901 18780447 142575529 161117193 108343354 121237253 17281185 5861482 120405281 ...
#>  $ geneSeqEnd    : int  108146146 77853623 18811987 142578143 161258213 108414396 121255856 17289115 5870394 120530186 ...
#>  $ seqName       : chr  "3" "X" "16" "7" ...
#>  $ seqStrand     : int  -1 -1 -1 -1 1 1 1 1 1 1 ...
#>  $ seqCoordSystem: chr  "chromosome" "chromosome" "chromosome" "chromosome" ...

# Alternate approach, with pre-compiled tibble

# NOT RUN {
annotable(annotables::grch37)
# }