Convert Ensembl Transcripts to Genes

Convert Ensembl Transcripts to Genes

convertTranscriptsToGenes(object, ...)

# S4 method for character
convertTranscriptsToGenes(object, tx2gene = NULL,
  organism = NULL, build = NULL, release = NULL)

# S4 method for matrix
convertTranscriptsToGenes(object, tx2gene = NULL,
  organism = NULL, build = NULL, release = NULL, aggregate = TRUE)

# S4 method for sparseMatrix
convertTranscriptsToGenes(object,
  tx2gene = NULL, organism = NULL, build = NULL, release = NULL,
  aggregate = TRUE)

# S4 method for SummarizedExperiment
convertTranscriptsToGenes(object)

Arguments

object
Object.

...
Passthrough to makeTx2geneFromEnsembl().

tx2gene
Tx2Gene or NULL. Transcript-to-gene mappings. If set NULL, the function will attempt to download the mappings from Ensembl automatically.

organism
string. Full latin organism name (e.g. "Homo sapiens").

build
string or NULL. Genome build assembly name (e.g. "GRCh38"). If set NULL, defaults to the most recent build available.

release
scalar integer or NULL. Release version (e.g. 90). If set NULL, defaults to the most recent release available.

aggregate
boolean. For objects supporting dim(), aggregate counts to gene level and collapse the matrix.

Value

character: factor. Genes in the values, transcripts in the names.

matrix, sparseMatrix, SummarizedExperiment: Object containing counts collapsed to gene level by default (see aggregate argument).

Note

For objects containing a count matrix, the object rows will be collapsed to gene level using aggregateRows(). This applies to our SummarizedExperiment method.

See also

Examples

data(tx_se_small)
object <- tx_se_small

tx2gene <- tx2gene(object)
print(tx2gene)

#> Tx2Gene with 6 rows and 2 columns
#>                    transcriptID          geneID
#>                     <character>     <character>
#> ENST00000494424 ENST00000494424 ENSG00000000003
#> ENST00000496771 ENST00000496771 ENSG00000000003
#> ENST00000612152 ENST00000612152 ENSG00000000003
#> ENST00000371584 ENST00000371584 ENSG00000000419
#> ENST00000371588 ENST00000371588 ENSG00000000419
#> ENST00000413082 ENST00000413082 ENSG00000000419

transcripts <- rownames(object)
print(transcripts)

#> [1] "ENST00000494424" "ENST00000496771" "ENST00000612152" "ENST00000371584"
#> [5] "ENST00000371588" "ENST00000413082"

# character ====
# Returns as factor.
x <- convertTranscriptsToGenes(transcripts, tx2gene = tx2gene)
print(x)

#> ENST00000494424 ENST00000496771 ENST00000612152 ENST00000371584 ENST00000371588 
#> ENSG00000000003 ENSG00000000003 ENSG00000000003 ENSG00000000419 ENSG00000000419 
#> ENST00000413082 
#> ENSG00000000419 
#> Levels: ENSG00000000003 ENSG00000000419

str(x)

#>  Factor w/ 2 levels "ENSG00000000003",..: 1 1 1 2 2 2
#>  - attr(*, "names")= chr [1:6] "ENST00000494424" "ENST00000496771" "ENST00000612152" "ENST00000371584" ...

# matrix ====
# Note that transcript IDs currently must be in the rows.
counts <- counts(object)
print(counts)

#>                 sample1 sample2 sample3 sample4
#> ENST00000494424       1       2       3       4
#> ENST00000496771       5       6       7       8
#> ENST00000612152       9      10      11      12
#> ENST00000371584      13      14      15      16
#> ENST00000371588      17      18      19      20
#> ENST00000413082      21      22      23      24

# Aggregate to gene level.
x <- convertTranscriptsToGenes(counts, tx2gene = tx2gene, aggregate = TRUE)

#> Aggregating counts using `sum()`...
#> Groupings:
#> ENST00000494424 ENST00000496771 ENST00000612152 ENST00000371584 ENST00000371588 
#> ENSG00000000003 ENSG00000000003 ENSG00000000003 ENSG00000000419 ENSG00000000419 
#> ENST00000413082 
#> ENSG00000000419 
#> Levels: ENSG00000000003 ENSG00000000419

print(x)

#>                 sample1 sample2 sample3 sample4
#> ENSG00000000003      15      18      21      24
#> ENSG00000000419      51      54      57      60

colSums(x)

#> sample1 sample2 sample3 sample4 
#>      66      72      78      84 

# Simply map to rownames.
x <- convertTranscriptsToGenes(counts, tx2gene = tx2gene, aggregate = FALSE)
print(x)

#>                 sample1 sample2 sample3 sample4
#> ENSG00000000003       1       2       3       4
#> ENSG00000000003       5       6       7       8
#> ENSG00000000003       9      10      11      12
#> ENSG00000000419      13      14      15      16
#> ENSG00000000419      17      18      19      20
#> ENSG00000000419      21      22      23      24

colSums(x)

#> sample1 sample2 sample3 sample4 
#>      66      72      78      84 

# SummarizedExperiment ====
x <- convertTranscriptsToGenes(object)

#> Aggregating counts using `sum()`...
#> Groupings:
#> ENST00000494424 ENST00000496771 ENST00000612152 ENST00000371584 ENST00000371588 
#> ENSG00000000003 ENSG00000000003 ENSG00000000003 ENSG00000000419 ENSG00000000419 
#> ENST00000413082 
#> ENSG00000000419 
#> Levels: ENSG00000000003 ENSG00000000419

print(x)

#> class: SummarizedExperiment 
#> dim: 2 4 
#> metadata(0):
#> assays(1): counts
#> rownames(2): ENSG00000000003 ENSG00000000419
#> rowData names(0):
#> colnames(4): sample1 sample2 sample3 sample4
#> colData names(0):