Convert Ensembl Transcripts to Genes
Source: R/AllGenerics.R, R/methods-convertTranscriptsToGenes.R

Convert Ensembl Transcripts to Genes

convertTranscriptsToGenes(object, ...)

# S4 method for character
convertTranscriptsToGenes(object, tx2gene = NULL,
  organism = NULL, genomeBuild = NULL, release = NULL)

# S4 method for matrix
convertTranscriptsToGenes(object, ...)

# S4 method for data.frame
convertTranscriptsToGenes(object, ...)

# S4 method for DataFrame
convertTranscriptsToGenes(object, ...)

# S4 method for dgCMatrix
convertTranscriptsToGenes(object, ...)

# S4 method for dgTMatrix
convertTranscriptsToGenes(object, ...)

Arguments

object
Object.

...
Additional arguments.

tx2gene
Optional. Transcript-to-gene mappings. If NULL, will attempt to download from Ensembl using the desired organism, genomeBuild, and release arguments.

organism
Optional. Organism name. Normally this argument is unnecessary and can be left unset. If a count matrix starts with a FASTA spike-in (e.g. "EGFP"), then automatic genome detection based on the first gene identifier will fail. In this case, the desired organism must be manually declared.

genomeBuild
Optional. Genome build assembly name (e.g. "GRCh38"). If NULL, defaults to the most recent build available.

release
Optional. Release version (e.g. 90). If NULL, defaults to the most recent release available.

Value

Same class as original object.

See also

Examples

# character ====
transcripts <- c(
    "ENSMUST00000000001",
    "ENSMUST00000000003",
    "ENSMUST00000114041"
)
convertTranscriptsToGenes(transcripts)

#> Obtaining transcript-to-gene mappings from Ensembl

#> Making GRanges from Ensembl with AnnotationHub 2.10.1 (2017-10-27)

#> id: "AH57770"
#> organism: "Mus musculus"
#> build: "GRCm38"
#> release: "90"
#> format: "transcripts"

#> Standardizing the metadata columns

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Setting txBiotype as factor

#> Setting txSupportLevel as factor

#> Defining broadClass using: geneName, txBiotype, seqnames

#>   ENSMUST00000000001   ENSMUST00000000003   ENSMUST00000114041 
#> "ENSMUSG00000000001" "ENSMUSG00000000003" "ENSMUSG00000000003" 

# matrix ====
mat <- matrix(
    data = seq(1L:6L),
    byrow = TRUE,
    nrow = 3L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUST00000000001", "ENSMUST00000000003", "ENSMUST00000114041"),
        c("sample_1", "sample_2")
    )
)
print(mat)

#>                    sample_1 sample_2
#> ENSMUST00000000001        1        2
#> ENSMUST00000000003        3        4
#> ENSMUST00000114041        5        6

mat <- convertTranscriptsToGenes(mat)

#> Obtaining transcript-to-gene mappings from Ensembl

#> Making GRanges from Ensembl with AnnotationHub 2.10.1 (2017-10-27)

#> id: "AH57770"
#> organism: "Mus musculus"
#> build: "GRCm38"
#> release: "90"
#> format: "transcripts"

#> Standardizing the metadata columns

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Setting txBiotype as factor

#> Setting txSupportLevel as factor

#> Defining broadClass using: geneName, txBiotype, seqnames

print(mat)

#>                    sample_1 sample_2
#> ENSMUSG00000000001        1        2
#> ENSMUSG00000000003        3        4
#> ENSMUSG00000000003        5        6

rownames(mat)

#>   ENSMUST00000000001   ENSMUST00000000003   ENSMUST00000114041 
#> "ENSMUSG00000000001" "ENSMUSG00000000003" "ENSMUSG00000000003"