Convert Ensembl Transcripts to Genes

Convert Ensembl Transcripts to Genes

convertTranscriptsToGenes(object, ...)

# S4 method for character
convertTranscriptsToGenes(object, tx2gene = NULL,
  organism = NULL, genomeBuild = NULL, release = NULL)

# S4 method for matrix
convertTranscriptsToGenes(object, ...)

# S4 method for data.frame
convertTranscriptsToGenes(object, ...)

# S4 method for DataFrame
convertTranscriptsToGenes(object, ...)

# S4 method for dgCMatrix
convertTranscriptsToGenes(object, ...)

# S4 method for dgTMatrix
convertTranscriptsToGenes(object, ...)

Arguments

object
Object.

...
Additional arguments.

tx2gene
data.frame or NULL. Transcript-to-gene mappings. If set NULL, the function will attempt to download mappings from Ensembl according to the organism, genomeBuild, and release parameters.

organism
string. Organism name. Normally this argument is unnecessary and can be left unset. If a count matrix starts with a FASTA spike-in (e.g. "EGFP"), then automatic genome detection based on the first gene identifier will fail. In this case, the desired organism must be manually declared.

genomeBuild
string or NULL. Genome build assembly name (e.g. "GRCh38"). If set NULL, defaults to the most recent build available.

release
scalar integer or NULL. Release version (e.g. 90). If set NULL, defaults to the most recent release available.

Value

Same class as original object.

See also

Examples

# character ====
x <- c("ENSMUST00000000001", "ENSMUST00000000003", "ENSMUST00000114041")
convertTranscriptsToGenes(x)

#> Obtaining transcript-to-gene mappings from Ensembl

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 92 EnsDb for Mus Musculus

#> id: "AH60992"
#> organism: "Mus musculus"
#> build: "GRCm38"
#> release: "92"
#> format: "transcripts"

#> Standardizing the metadata columns

#> Defining broad class using: geneName, transcriptBiotype, seqnames

#> Sorting ranges by transcriptID

#>   ENSMUST00000000001   ENSMUST00000000003   ENSMUST00000114041 
#> "ENSMUSG00000000001" "ENSMUSG00000000003" "ENSMUSG00000000003" 

# matrix ====
mat <- matrix(
    data = seq(1L:6L),
    byrow = TRUE,
    nrow = 3L,
    ncol = 2L,
    dimnames = list(
        c("ENSMUST00000000001", "ENSMUST00000000003", "ENSMUST00000114041"),
        c("sample_1", "sample_2")
    )
)
print(mat)

#>                    sample_1 sample_2
#> ENSMUST00000000001        1        2
#> ENSMUST00000000003        3        4
#> ENSMUST00000114041        5        6

mat <- convertTranscriptsToGenes(mat)

#> Obtaining transcript-to-gene mappings from Ensembl

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 92 EnsDb for Mus Musculus

#> id: "AH60992"
#> organism: "Mus musculus"
#> build: "GRCm38"
#> release: "92"
#> format: "transcripts"

#> Standardizing the metadata columns

#> Defining broad class using: geneName, transcriptBiotype, seqnames

#> Sorting ranges by transcriptID

print(mat)

#>                    sample_1 sample_2
#> ENSMUSG00000000001        1        2
#> ENSMUSG00000000003        3        4
#> ENSMUSG00000000003        5        6

rownames(mat)

#>   ENSMUST00000000001   ENSMUST00000000003   ENSMUST00000114041 
#> "ENSMUSG00000000001" "ENSMUSG00000000003" "ENSMUSG00000000003"