Ensembl Annotations

Quickly obtain gene and transcript annotations from Ensembl using AnnotationHub.

ensembl(organism, format = c("genes", "gene2symbol", "transcripts",
  "tx2gene"), genomeBuild = NULL, release = NULL, metadata = FALSE,
  return = c("GRanges", "DataFrame", "data.frame"))

Arguments

organism
Default recommended usage is to provide the full latin organism name as a string.

format
Desired annotation data format, either "genes", "transcripts", "gene2symbol"., or "tx2gene".

genomeBuild
Optional. Genome assembly name (e.g. GRCh38).

release
Optional. Ensembl release version. Defaults to the most current release available on AnnotationHub.

metadata
Include the AnnotationHub metadata inside a list. Useful for documenting where the annotations were sourced from inside AnnotationHub.

return
Class of the returned object. Only applies when format equals "genes" or "transcripts". Can be "GRanges" (default), "DataFrame", or "data.frame". See help("genes", "ensembldb") for additional information. "gene2symbol" and "tx2gene" are always returned as data.frame.

Value

GRanges, DataFrame, or data.frame containing gene or transcript annotations.

Details

Simply specify the desired organism, using the full latin name. For example, we can obtain human annotations with Homo sapiens. Optionally, specific Ensembl genome builds (e.g. GRCh38) and release versions (e.g. 87) are supported.

Under the hood, this function fetches annotations from AnnotationHub using the ensembldb package. AnnotationHub supports versioned Ensembl releases, back to version 87.

Note

Use GRCh38 instead of hg38 for the genome build, since we're querying Ensembl and not UCSC. Unfortunately, GRCh37 is not currently suported on AnnotationHub.

Broad Class Definitions

For gene and transcript tables, when broadClass = TRUE, a broadClass column is added, which generalizes the gene types into a smaller number of semantically-meaningful groups:

coding

noncoding

pseudo

small

decaying

ig (immunoglobulin)

tcr (T cell receptor)

other

See also

AnnotationHub.

ensembldb.

Other Gene Functions: convertGenesToSymbols, convertTranscriptsToGenes, detectOrganism, gene2symbolFromGFF, gene2symbol, stripTranscriptVersions, tx2geneFromGFF, tx2gene

Examples

# Genes ====
data <- ensembl("Homo sapiens", format = "genes")

#> Fetching Ensembl genes from AnnotationHub 2.10.1 (2017-10-27)

#> hub id: "AH57757"
#> organism: "Homo sapiens"
#> genome build: "GRCh38"
#> release version: "90"

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

summary(data)

#> [1] "GRanges object with 64661 ranges and 7 metadata columns"

# Transcripts ====
data <- ensembl("Homo sapiens", format = "transcripts")

#> Fetching Ensembl transcripts from AnnotationHub 2.10.1 (2017-10-27)

#> hub id: "AH57757"
#> organism: "Homo sapiens"
#> genome build: "GRCh38"
#> release version: "90"

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Setting txBiotype as factor

#> Setting txSupportLevel as factor

#> Defining broadClass using: geneName, txBiotype, seqnames

summary(data)

#> [1] "GRanges object with 220144 ranges and 13 metadata columns"