Return the first and last parts of an object

Applies to both rows and columns, enabling quick inspection during interactive use.

headtail(x, ...)

# S4 method for DataFrame
headtail(x, n = 2L)

# S4 method for Matrix
headtail(x, n = 2L)

# S4 method for GRanges
headtail(x, n = 2L)

# S4 method for SummarizedExperiment
headtail(x, n = 2L)

Arguments

x
Object.

n
integer(1). Positive integer denoting the number of first and last items to include.

...
Additional arguments.

Value

Invisible NULL.

Details

Inspired by the print() method for DataFrame class objects.

Functions

headtail,DataFrame-method: Same method as data.frame.

headtail,Matrix-method: Same method as matrix.

headtail,GRanges-method: Summarize the ranges.

headtail,SummarizedExperiment-method: Summarize the primary assay.

Note

Updated 2020-10-07.

See also

Examples

data(mtcars, package = "datasets")
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
headtail(rse)

#>         sample01 sample02 ... sample11 sample12
#> gene001        2        7 ...        1        2
#> gene002       17        7 ...       28       18
#> ...          ...      ... ...      ...      ...
#> gene499        1        0 ...        0        0
#> gene500       27       13 ...       12       36