Return the First and Last Part of an Object

Inspired by the print() method for DataFrame class objects. Applies to both rows and columns, enabling quick inspection during interactive use.

headtail(x, ...)

# S4 method for atomic
headtail(x, n = 2L)

# S4 method for matrix
headtail(x, n = 2L)

# S4 method for sparseMatrix
headtail(x, n = 2L)

# S4 method for data.frame
headtail(x, n = 2L)

# S4 method for DataFrame
headtail(x, n = 2L)

# S4 method for GRanges
headtail(x, n = 2L)

# S4 method for SummarizedExperiment
headtail(x, n = 2L)

Arguments

x
Object.

...
Additional arguments.

n
scalar integer. Positive integer denoting the number of first and last items to include.

Value

None (invisible NULL).

Methods (by class)

atomic: Paste collapse to a string.

matrix: Show first and last rows.

sparseMatrix: Same method as matrix.

data.frame: Same method as matrix.

DataFrame: Same method as data.frame.

GRanges: Summarize the ranges.

SummarizedExperiment: Summarize the primary assay().

See also

Examples

data(rse_small, sce_small)

## data.frame ====
headtail(datasets::mtcars)

#>                mpg cyl ... gear carb
#> Mazda RX4       21   6 ...    4    4
#> Mazda RX4 Wag   21   6 ...    4    4
#> ...            ... ... ...  ...  ...
#> Maserati Bora   15   8 ...    5    8
#> Volvo 142E    21.4   4 ...    4    2

## SummarizedExperiment ====
headtail(rse_small)

#>        sample1 sample2 ... sample3 sample4
#> gene1        7       4 ...      14      50
#> gene2       24       2 ...       4       1
#> ...        ...     ... ...     ...     ...
#> gene49       1       1 ...       1       2
#> gene50       7      14 ...      14       9

headtail(colData(rse_small))

#> Error in .local(x, ...): ncol(x) >= n * 2L is not TRUE

headtail(rowData(rse_small))

#>                 geneID geneName ... broadClass entrezID
#> gene1  ENSG00000000003   TSPAN6 ...     coding     7105
#> gene2  ENSG00000000005     TNMD ...     coding    64102
#> ...                ...      ... ...        ...      ...
#> gene49 ENSG00000004399   PLXND1 ...     coding    23129
#> gene50 ENSG00000004455      AK2 ...     coding      204

## SingleCellExperiment ====
headtail(sce_small)

#>                 cell001 cell002 ... cell999 cell1000
#> ENSG00000000003      82      52 ...      48       93
#> ENSG00000000005       0      99 ...       0      138
#> ...                 ...     ... ...     ...      ...
#> ENSG00000008056     423     305 ...     391      487
#> ENSG00000008083     133     101 ...     133      168