Import

Read file by extension into R.

import(file, ...)

Arguments

file
string. File path.

...
Additional arguments.

Value

Varies, depending on the file extension:

Default: DataFrame.

MTX: sparseMatrix.

GTF/GFF: GRanges.

JSON/YAML: list.

Note that data frames are returned as S4 DataFrame class instead of data.frame, when applicable.

Details

This is a wrapper for rio::import() that adds support for additional common genomic data files. Remote URLs and compressed files are supported.

This function supports automatic loading of common file types:

csv: Comma Separated Values.

tsv Tab Separated Values.

xlsx: Excel workbook.

mtx: MatrixMarket sparse matrix.

gtf/gff/gff3: General Feature Format.

rda/RData: R Data. Must contain a single object. Doesn't require internal object name to match, like with the loadData() function.

rds: R Data Serialized.

json: JSON.

yaml/yml: YAML.

If the file format isn't supported natively (or blacklisted), the rio package will be used as a fallback attempt.

Matrix Market Exchange (MTX/MEX)

Reading a Matrix Market Exchange ("mtx") file now requires "colnames" and "rownames" sidecar files containing the colnames() and rownames() of the sparse matrix. Legacy support for manual loading of these sidecar files is provided.

General Feature Format (GFF/GTF)

The GFF (General Feature Format) format consists of one line per feature, each containing 9 columns of data, plus optional track definition lines. The GTF (General Transfer Format) is identical to GFF version 2.

Column names follow the Ensembl conventions. Additional information:

Ensembl

Gencode

bcbio

Also supports some additional extensions commonly used with the bcbio pipeline:

counts: Counts table (e.g. RNA-seq aligned counts).

colnames: Sidecar file containing column names.

rownames: Sidecar file containing row names.

See also

Examples

# R Data
x <- import(file = file.path(basejumpCacheURL, "rnaseq_counts.rda"))

#> Reading rnaseq_counts.rda...

class(x)

#> [1] "matrix"

# Comma Separated Values
x <- import(file = file.path(basejumpCacheURL, "example.csv"))

#> Reading example.csv...

print(x)

#> DataFrame with 4 rows and 2 columns
#>            genotype   treatment
#>         <character> <character>
#> sample1    wildtype     control
#> sample2    knockout     control
#> sample3    wildtype     treated
#> sample4    knockout     treated

# Microsoft Excel Worksheet
x <- import(file = file.path(basejumpCacheURL, "example.xlsx"))

#> Reading example.xlsx...

print(x)

#> DataFrame with 6 rows and 11 columns
#>                         mpg       cyl      disp        hp      drat        wt
#>                   <numeric> <numeric> <numeric> <numeric> <numeric> <numeric>
#> Mazda RX4                21         6       160       110       3.9      2.62
#> Mazda RX4 Wag            21         6       160       110       3.9     2.875
#> Datsun 710             22.8         4       108        93      3.85      2.32
#> Hornet 4 Drive         21.4         6       258       110      3.08     3.215
#> Hornet Sportabout      18.7         8       360       175      3.15      3.44
#> Valiant                18.1         6       225       105      2.76      3.46
#>                        qsec        vs        am      gear      carb
#>                   <numeric> <numeric> <numeric> <numeric> <numeric>
#> Mazda RX4             16.46         0         1         4         4
#> Mazda RX4 Wag         17.02         0         1         4         4
#> Datsun 710            18.61         1         1         4         1
#> Hornet 4 Drive        19.44         1         0         3         1
#> Hornet Sportabout     17.02         0         0         3         2
#> Valiant               20.22         1         0         3         1

# GTF/GFF
x <- import(file = file.path(basejumpCacheURL, "example.gtf"))

#> Reading example.gtf...

summary(x)

#> [1] "GRanges object with 95 ranges and 22 metadata columns"

# JSON
x <- import(file = file.path(basejumpCacheURL, "example.json"))

#> Reading example.json...

names(x)

#> [1] "fasta_hash"        "genomes"           "gtf_hash"         
#> [4] "input_fasta_files" "input_gtf_files"   "mem_gb"           
#> [7] "mkref_version"     "threads"           "version"          

# YAML
x <- import(file = file.path(basejumpCacheURL, "example.yml"))

#> Reading example.yml...

names(x)

#> [1] "date"         "upload"       "bcbio_system" "samples"     

# Counts Table (i.e. aligned counts from bcbio)
x <- import(file = file.path(basejumpCacheURL, "example.counts"))

#> Reading example.counts...

#> Parsed with column specification:
#> cols(
#>   id = col_character(),
#>   sample01 = col_integer(),
#>   sample02 = col_integer(),
#>   sample03 = col_integer(),
#>   sample04 = col_integer(),
#>   sample05 = col_integer(),
#>   sample06 = col_integer(),
#>   sample07 = col_integer(),
#>   sample08 = col_integer(),
#>   sample09 = col_integer(),
#>   sample10 = col_integer(),
#>   sample11 = col_integer(),
#>   sample12 = col_integer()
#> )

colSums(x)

#> sample01 sample02 sample03 sample04 sample05 sample06 sample07 sample08 
#>      262      502      299      318      123      620      242      303 
#> sample09 sample10 sample11 sample12 
#>      199      258      258      157