Genomic Ranges from Ensembl

Quickly obtain gene and transcript annotations from Ensembl using AnnotationHub and ensembldb.

makeGRangesFromEnsembl(organism, format = c("genes", "transcripts"),
  genomeBuild = NULL, release = NULL, metadata = FALSE)

Arguments

organism
string. Full latin organism name (e.g. "Homo sapiens").

format
Fetch ranges as "genes" or "transcripts".

genomeBuild
Optional. Genome build assembly name (e.g. "GRCh38"). If NULL, defaults to the most recent build available.

release
Optional. Release version (e.g. 90). If NULL, defaults to the most recent release available.

metadata
Include the AnnotationHub metadata inside a list.

Value

GRanges.

Details

Simply specify the desired organism, using the full latin name. For example, we can obtain human annotations with Homo sapiens. Optionally, specific Ensembl genome builds (e.g. GRCh38) and release versions (e.g. 87) are supported.

Under the hood, this function fetches annotations from AnnotationHub using the ensembldb package. AnnotationHub supports versioned Ensembl releases, back to version 87.

Note

Use "GRCh38" instead of "hg38" for the genome build, since we're querying Ensembl and not UCSC.

Broad class definitions

For gene and transcript annotations, a broadClass column is added, which generalizes the gene types into a smaller number of semantically-meaningful groups:

coding

noncoding

pseudo

small

decaying

ig (immunoglobulin)

tcr (T cell receptor)

other

GRCh37 legacy annotations

The Homo sapiens GRCh37 (release 75) build is supported by internally querying the EnsDb.Hsapiens.v75 package.

See also

Examples

# Genes ====
x <- makeGRangesFromEnsembl("Homo sapiens", format = "genes")

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 92 EnsDb for Homo Sapiens

#> id: "AH60977"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "92"
#> format: "genes"

#> Standardizing the metadata columns

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

summary(x)

#> [1] "GRanges object with 65256 ranges and 7 metadata columns"

# Transcripts ====
x <- makeGRangesFromEnsembl("Homo sapiens", format = "transcripts")

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 92 EnsDb for Homo Sapiens

#> id: "AH60977"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "92"
#> format: "transcripts"

#> Standardizing the metadata columns

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Setting txBiotype as factor

#> Setting txSupportLevel as factor

#> Defining broadClass using: geneName, txBiotype, seqnames

summary(x)

#> [1] "GRanges object with 225358 ranges and 13 metadata columns"