Genomic Ranges from GFF File

The GFF (General Feature Format) format consists of one line per feature, each containing 9 columns of data, plus optional track definition lines. The GTF (General Transfer Format) is identical to GFF version 2.

makeGRangesFromGFF(file, format = c("genes", "transcripts"))

Arguments

file
File path.

format
Output as genes or transcripts.

Value

GRanges.

See also

Examples

file <- "http://basejump.seq.cloud/mmusculus.gtf"

# Genes
invisible(capture.output(
    x <- makeGRangesFromGFF(file, format = "genes")
))

#> Reading mmusculus.gtf

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> OK

#> Standardizing the metadata columns

#> Setting geneBiotype as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

summary(x)

#> [1] "GRanges object with 17 ranges and 6 metadata columns"

# Transcripts
invisible(capture.output(
    x <- makeGRangesFromGFF(file, format = "transcripts")
))

#> Reading mmusculus.gtf

#> Import genomic features from the file as a GRanges object ... 

#> OK

#> Prepare the 'metadata' data frame ... 

#> OK

#> Make the TxDb object ... 

#> OK

#> Standardizing the metadata columns

#> Setting geneBiotype as factor

#> Setting txBiotype as factor

#> Defining broadClass using: geneName, txBiotype, seqnames

summary(x)

#> [1] "GRanges object with 20 ranges and 12 metadata columns"