Make Syntactically Valid Names

Make Syntactically Valid Names

makeNames(names, unique = FALSE)

Arguments

names
character vector to be coerced to syntactically valid names. This is coerced to character if necessary.

unique
logical; if TRUE, the resulting elements are unique. This may be desired for, e.g., column names.

Value

Character vector.

Details

A syntactically valid name consists of letters, numbers and the dot or underline characters and starts with a letter or the dot not followed by a number. Names such as ".2way" are not valid, and neither are the reserved words.

The definition of a letter depends on the current locale, but only ASCII digits are considered to be digits.

The character "X" is prepended if necessary. All invalid characters are translated to ".". A missing value is translated to "NA". Names which match R keywords have a dot appended to them. Duplicated values are altered by make.unique.

Note

Sanitizes names using "_" instead of ".".

Warning

Some OSes, notably FreeBSD, report extremely incorrect information about which characters are alphabetic in some locales (typically, all multi-byte locales including UTF-8 locales). However, R provides substitutes on Windows, macOS and AIX.

See also

make.names().