Make Syntactically Valid Names

For atomic vectors, these functions will sanitize the values. Otherwise, they will set names(), rownames(), and/or colnames() without modification of the values.

camel(object, ...)

dotted(object, ...)

snake(object, ...)

upperCamel(object, ...)

makeNames(names, unique = FALSE)

# S4 method for character
camel(object, strict = FALSE)

# S4 method for character
dotted(object)

# S4 method for character
snake(object)

# S4 method for character
upperCamel(object, strict = FALSE)

# S4 method for factor
camel(object, strict = FALSE)

# S4 method for factor
dotted(object)

# S4 method for factor
snake(object)

# S4 method for factor
upperCamel(object, strict = FALSE)

# S4 method for matrix
camel(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

# S4 method for matrix
dotted(object, rownames = FALSE, colnames = TRUE)

# S4 method for matrix
snake(object, rownames = FALSE, colnames = TRUE)

# S4 method for matrix
upperCamel(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

# S4 method for data.frame
camel(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

# S4 method for data.frame
dotted(object, rownames = FALSE,
  colnames = TRUE)

# S4 method for data.frame
snake(object, rownames = FALSE, colnames = TRUE)

# S4 method for data.frame
upperCamel(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

# S4 method for DataFrame
camel(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

# S4 method for DataFrame
dotted(object, rownames = FALSE, colnames = TRUE)

# S4 method for DataFrame
snake(object, rownames = FALSE, colnames = TRUE)

# S4 method for DataFrame
upperCamel(object, rownames = FALSE,
  colnames = TRUE, strict = FALSE)

# S4 method for GRanges
camel(object, strict = FALSE)

# S4 method for GRanges
dotted(object)

# S4 method for GRanges
snake(object)

# S4 method for GRanges
upperCamel(object, strict = FALSE)

# S4 method for list
camel(object, strict = FALSE)

# S4 method for list
dotted(object)

# S4 method for list
snake(object)

# S4 method for list
upperCamel(object, strict = FALSE)

# S4 method for List
camel(object, strict = FALSE)

# S4 method for List
dotted(object)

# S4 method for List
snake(object)

# S4 method for List
upperCamel(object, strict = FALSE)

# S4 method for SimpleList
camel(object, strict = FALSE)

# S4 method for SimpleList
dotted(object)

# S4 method for SimpleList
snake(object)

# S4 method for SimpleList
upperCamel(object, strict = FALSE)

# S4 method for ANY
camel(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

# S4 method for ANY
dotted(object, rownames = FALSE, colnames = TRUE)

# S4 method for ANY
snake(object, rownames = FALSE, colnames = TRUE)

# S4 method for ANY
upperCamel(object, rownames = FALSE, colnames = TRUE,
  strict = FALSE)

Arguments

object
Character vector or an object for which names() assignment will be meaningful.

...
Additional arguments.

names
atomic to be coerced to syntactically valid names. Will be coerced to character, if necessary.

unique
boolean. If TRUE, the resulting elements are unique.

strict
boolean. Enforce strict name sanitization. When TRUE, this does not allow the return of any capitalized acronyms. "RNA" will become "Rna", for example.

rownames
boolean. Apply sanitization on row names. This is not recommended by default, since rownames commonly contain gene identifiers that should not be modified.

colnames
boolean. Apply sanitization on column names. This is generally recommended by default.

Value

Object with syntatically valid names. For objects supporting names(), the underlying data returns unchanged.

Note

dotted.case support is provided for matching against base R parameters. However, it is recommended to avoid using it for variable assignments into an environment, as that can introduce conflicts with base functions.

makeNames() sanitizes names using underscore instead of dot, which is the convention used by make.names().

See also

Examples

loadRemoteData("http://basejump.seq.cloud/mn.rda")

#> Reading mn.rda

# character ====
x <- mn$character
print(x)

#>  [1] "hello world"     "HELLO WORLD"     "RNAi clones"     "nCount"         
#>  [5] "tx2gene"         "TX2GeneID"       "G2M.Score"       "worfdbHTMLRemap"
#>  [9] "Mazda RX4"       "123"             NA               

camel(x)

#>  [1] "helloWorld"      "helloWORLD"      "rnaiClones"      "nCount"         
#>  [5] "tx2gene"         "tx2GeneID"       "g2mScore"        "worfdbHTMLRemap"
#>  [9] "mazdaRX4"        "x123"            NA               

dotted(x)

#>  [1] "hello.world"       "HELLO.WORLD"       "RNAI.clones"      
#>  [4] "n.Count"           "tx2gene"           "TX2.Gene.ID"      
#>  [7] "G2M.Score"         "worfdb.HTML.Remap" "Mazda.RX4"        
#> [10] "X123"              NA                 

snake(x)

#>  [1] "hello_world"       "hello_world"       "rnai_clones"      
#>  [4] "n_count"           "tx2gene"           "tx2_gene_id"      
#>  [7] "g2m_score"         "worfdb_html_remap" "mazda_rx4"        
#> [10] "x123"              NA                 

upperCamel(x)

#>  [1] "HelloWorld"      "HELLOWORLD"      "RNAIClones"      "NCount"         
#>  [5] "Tx2gene"         "TX2GeneID"       "G2MScore"        "WorfdbHTMLRemap"
#>  [9] "MazdaRX4"        "X123"            NA               

makeNames(x)

#>  [1] "hello_world"     "HELLO_WORLD"     "RNAi_clones"     "nCount"         
#>  [5] "tx2gene"         "TX2GeneID"       "G2M_Score"       "worfdbHTMLRemap"
#>  [9] "Mazda_RX4"       "X123"            "NA_"            

x <- mn$namedCharacter
print(x)

#>        Item.A        Item.B 
#> "hello world" "HELLO WORLD" 

camel(x)

#>        itemA        itemB 
#> "helloWorld" "helloWORLD" 

dotted(x)

#>        Item.A        Item.B 
#> "hello.world" "HELLO.WORLD" 

snake(x)

#>        item_a        item_b 
#> "hello_world" "hello_world" 

upperCamel(x)

#>        ItemA        ItemB 
#> "HelloWorld" "HELLOWORLD" 

makeNames(x)

#> [1] "hello_world" "HELLO_WORLD"

# factor ====
x <- mn$factor
print(x)

#> sample 1 sample 2 sample 3 sample 4 
#>  group 1  group 1  group 2  group 2 
#> Levels: group 1 group 2

camel(x)

#> sample1 sample2 sample3 sample4 
#>  group1  group1  group2  group2 
#> Levels: group1 group2

dotted(x)

#> sample.1 sample.2 sample.3 sample.4 
#>  group.1  group.1  group.2  group.2 
#> Levels: group.1 group.2

snake(x)

#> sample_1 sample_2 sample_3 sample_4 
#>  group_1  group_1  group_2  group_2 
#> Levels: group_1 group_2

upperCamel(x)

#> Sample1 Sample2 Sample3 Sample4 
#>  Group1  Group1  Group2  Group2 
#> Levels: Group1 Group2

makeNames(x)

#> [1] "group_1" "group_1" "group_2" "group_2"

# data.frame ====
x <- datasets::USArrests
dimnames(x)

#> [[1]]
#>  [1] "Alabama"        "Alaska"         "Arizona"        "Arkansas"      
#>  [5] "California"     "Colorado"       "Connecticut"    "Delaware"      
#>  [9] "Florida"        "Georgia"        "Hawaii"         "Idaho"         
#> [13] "Illinois"       "Indiana"        "Iowa"           "Kansas"        
#> [17] "Kentucky"       "Louisiana"      "Maine"          "Maryland"      
#> [21] "Massachusetts"  "Michigan"       "Minnesota"      "Mississippi"   
#> [25] "Missouri"       "Montana"        "Nebraska"       "Nevada"        
#> [29] "New Hampshire"  "New Jersey"     "New Mexico"     "New York"      
#> [33] "North Carolina" "North Dakota"   "Ohio"           "Oklahoma"      
#> [37] "Oregon"         "Pennsylvania"   "Rhode Island"   "South Carolina"
#> [41] "South Dakota"   "Tennessee"      "Texas"          "Utah"          
#> [45] "Vermont"        "Virginia"       "Washington"     "West Virginia" 
#> [49] "Wisconsin"      "Wyoming"       
#> 
#> [[2]]
#> [1] "Murder"   "Assault"  "UrbanPop" "Rape"    
#> 

camel(x, rownames = TRUE, colnames = TRUE) %>% dimnames()

#> [[1]]
#>  [1] "alabama"       "alaska"        "arizona"       "arkansas"     
#>  [5] "california"    "colorado"      "connecticut"   "delaware"     
#>  [9] "florida"       "georgia"       "hawaii"        "idaho"        
#> [13] "illinois"      "indiana"       "iowa"          "kansas"       
#> [17] "kentucky"      "louisiana"     "maine"         "maryland"     
#> [21] "massachusetts" "michigan"      "minnesota"     "mississippi"  
#> [25] "missouri"      "montana"       "nebraska"      "nevada"       
#> [29] "newHampshire"  "newJersey"     "newMexico"     "newYork"      
#> [33] "northCarolina" "northDakota"   "ohio"          "oklahoma"     
#> [37] "oregon"        "pennsylvania"  "rhodeIsland"   "southCarolina"
#> [41] "southDakota"   "tennessee"     "texas"         "utah"         
#> [45] "vermont"       "virginia"      "washington"    "westVirginia" 
#> [49] "wisconsin"     "wyoming"      
#> 
#> [[2]]
#> [1] "murder"   "assault"  "urbanPop" "rape"    
#> 

dotted(x, rownames = TRUE, colnames = TRUE) %>% dimnames()

#> [[1]]
#>  [1] "Alabama"        "Alaska"         "Arizona"        "Arkansas"      
#>  [5] "California"     "Colorado"       "Connecticut"    "Delaware"      
#>  [9] "Florida"        "Georgia"        "Hawaii"         "Idaho"         
#> [13] "Illinois"       "Indiana"        "Iowa"           "Kansas"        
#> [17] "Kentucky"       "Louisiana"      "Maine"          "Maryland"      
#> [21] "Massachusetts"  "Michigan"       "Minnesota"      "Mississippi"   
#> [25] "Missouri"       "Montana"        "Nebraska"       "Nevada"        
#> [29] "New.Hampshire"  "New.Jersey"     "New.Mexico"     "New.York"      
#> [33] "North.Carolina" "North.Dakota"   "Ohio"           "Oklahoma"      
#> [37] "Oregon"         "Pennsylvania"   "Rhode.Island"   "South.Carolina"
#> [41] "South.Dakota"   "Tennessee"      "Texas"          "Utah"          
#> [45] "Vermont"        "Virginia"       "Washington"     "West.Virginia" 
#> [49] "Wisconsin"      "Wyoming"       
#> 
#> [[2]]
#> [1] "Murder"    "Assault"   "Urban.Pop" "Rape"     
#> 

snake(x, rownames = TRUE, colnames = TRUE) %>% dimnames()

#> [[1]]
#>  [1] "alabama"        "alaska"         "arizona"        "arkansas"      
#>  [5] "california"     "colorado"       "connecticut"    "delaware"      
#>  [9] "florida"        "georgia"        "hawaii"         "idaho"         
#> [13] "illinois"       "indiana"        "iowa"           "kansas"        
#> [17] "kentucky"       "louisiana"      "maine"          "maryland"      
#> [21] "massachusetts"  "michigan"       "minnesota"      "mississippi"   
#> [25] "missouri"       "montana"        "nebraska"       "nevada"        
#> [29] "new_hampshire"  "new_jersey"     "new_mexico"     "new_york"      
#> [33] "north_carolina" "north_dakota"   "ohio"           "oklahoma"      
#> [37] "oregon"         "pennsylvania"   "rhode_island"   "south_carolina"
#> [41] "south_dakota"   "tennessee"      "texas"          "utah"          
#> [45] "vermont"        "virginia"       "washington"     "west_virginia" 
#> [49] "wisconsin"      "wyoming"       
#> 
#> [[2]]
#> [1] "murder"    "assault"   "urban_pop" "rape"     
#> 

upperCamel(x, rownames = TRUE, colnames = TRUE) %>% dimnames()

#> [[1]]
#>  [1] "Alabama"       "Alaska"        "Arizona"       "Arkansas"     
#>  [5] "California"    "Colorado"      "Connecticut"   "Delaware"     
#>  [9] "Florida"       "Georgia"       "Hawaii"        "Idaho"        
#> [13] "Illinois"      "Indiana"       "Iowa"          "Kansas"       
#> [17] "Kentucky"      "Louisiana"     "Maine"         "Maryland"     
#> [21] "Massachusetts" "Michigan"      "Minnesota"     "Mississippi"  
#> [25] "Missouri"      "Montana"       "Nebraska"      "Nevada"       
#> [29] "NewHampshire"  "NewJersey"     "NewMexico"     "NewYork"      
#> [33] "NorthCarolina" "NorthDakota"   "Ohio"          "Oklahoma"     
#> [37] "Oregon"        "Pennsylvania"  "RhodeIsland"   "SouthCarolina"
#> [41] "SouthDakota"   "Tennessee"     "Texas"         "Utah"         
#> [45] "Vermont"       "Virginia"      "Washington"    "WestVirginia" 
#> [49] "Wisconsin"     "Wyoming"      
#> 
#> [[2]]
#> [1] "Murder"   "Assault"  "UrbanPop" "Rape"    
#> 

# list ====
x <- mn$list
print(x)

#> $Item.A
#> [1] 1 2
#> 
#> $Item.B
#> [1] 3 4
#> 

camel(x) %>% names()

#> [1] "itemA" "itemB"

dotted(x) %>% names()

#> [1] "Item.A" "Item.B"

snake(x) %>% names()

#> [1] "item_a" "item_b"

upperCamel(x) %>% names()

#> [1] "ItemA" "ItemB"