Make Summarized Experiment

This function is a utility wrapper for SummarizedExperiment::SummarizedExperiment() that provides automatic subsetting for row and column data, as well as automatic handling of transgenes and spike-ins. Additionally, it improves upon the standard constructor by slotting useful session information into the metadata() slot:

makeSummarizedExperiment(assays, rowRanges = NULL, colData = NULL,
  metadata = NULL, transgeneNames = NULL, spikeNames = NULL)

Arguments

assays
list. RNA-seq count matrices, which must have matching dimensions. Counts can be passed in either dense (matrix) or sparse (dgCMatrix, dgTMatrix) format.

rowRanges
GRanges or NULL. Metadata describing the assay rows. If defined, must contain genomic ranges. Can be left NULL if the genome is poorly annotated and/or ranges aren't available from AnnotationHub.

colData
DataFrame data.frame, or matrix. Metadata describing the assay columns. For bulk RNA-seq, this data describes the samples. For single-cell RNA-seq, this data describes the cells.

metadata
list or NULL. Metadata.

transgeneNames
character or NULL. Vector indicating which assay() rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character or NULL. Vector indicating which assay() rows denote spike-in sequences (e.g. ERCCs).

Value

RangedSummarizedExperiment.

Details

date: Today's date, returned from base::Sys.Date().

wd: Working directory, returned from base::getwd().

utilsSessionInfo: utils::sessionInfo().

devtoolsSessionInfo: sessioninfo::session_info().

See also

Examples

genes <- c(
    "EGFP",  # transgene
    "gene1",
    "gene2",
    "gene3"
)
samples <- c(
    "sample1",
    "sample2",
    "sample3",
    "sample4"
)

# Example matrix
mat <- matrix(
    seq(1L:16L),
    nrow = 4L,
    ncol = 4L,
    dimnames = list(genes, samples)
)

# Primary assay must be named "counts"
assays = list(counts = mat)

# rowRanges won't contain transgenes or spike-ins
rowRanges <- GRanges(
    seqnames = c("1", "1", "1"),
    ranges = IRanges(
        start = c(1L, 101L, 201L),
        end = c(100L, 200L, 300L)
    )
)
names(rowRanges) <- c("gene1", "gene2", "gene3")

colData <- data.frame(
    genotype = c(
        "wildtype",
        "wildtype",
        "knockout",
        "knockout"
    ),
    age = c(3L, 6L, 3L, 6L),
    row.names = samples
)

makeSummarizedExperiment(
    assays = assays,
    rowRanges = rowRanges,
    colData = colData,
    transgeneNames = "EGFP"
)

#> class: RangedSummarizedExperiment 
#> dim: 4 4 
#> metadata(4): date wd utilsSessionInfo devtoolsSessionInfo
#> assays(1): counts
#> rownames(4): EGFP gene1 gene2 gene3
#> rowData names(0):
#> colnames(4): sample1 sample2 sample3 sample4
#> colData names(2): genotype age