Make syntactically valid names

These are convenience functions that sanitize names into camelCase (preferred), snake_case, or dot.notation. Dot assignment is provided for use with base R operations, but we advise against using dot.notation name assignments for values and/or function.

camel(object, rownames = TRUE)

dotted(object, rownames = TRUE)

snake(object, rownames = TRUE)

titleCase(object)

firstCase(object)

Arguments

object
Character vector or an object for which names() assignment will be meaningful.

rownames
Include row names.

Value

Object with syntatically valid names. For objects supporting base::names(), the underlying data returns unchanged.

Details

For unnamed character vectors, these functions will sanitize the underlying values. Otherwise, the functions will set names() and/or rownames() on objects supporting name assignments. They return the object without modification of the underlying data.

Examples

# Unnamed character vector, with no names assigned
unnamed_vec <- c("hello world",
                 "HELLO WORLD",
                 "RNAi clones",
                 "worfdbHTMLRemap",
                 123,
                 NA)

# Named character vector
named_vec <- c(Item.A = "hello world", Item.B = "HELLO WORLD")

# Data frame with colnames and rownames
df <- head(mtcars)

# Tibble, with colnames. Rownames aren't supported on these!
tbl <- head(starwars)

# List, with assigned namess
lst <- list(Item.A = c(1, 2),
            Item.B = c(3, 4))

# camelCase
camel(unnamed_vec)

#> [1] "helloWorld"      "helloWorld"      "rnaiClones"      "worfdbHtmlRemap"
#> [5] "x123"            "na"             

camel(named_vec)

#>         itemA         itemB 
#> "hello world" "HELLO WORLD" 

camel(df)

#>                   mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> mazdaRx4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> mazdaRx4Wag      21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> datsun710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> hornet4Drive     21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> hornetSportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> valiant          18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

camel(df, rownames = FALSE)

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

camel(tbl)

#> # A tibble: 6 x 13
#>             name height  mass   hairColor   skinColor eyeColor birthYear gender
#>            <chr>  <int> <dbl>       <chr>       <chr>    <chr>     <dbl>  <chr>
#> 1 Luke Skywalker    172    77       blond        fair     blue      19.0   male
#> 2          C-3PO    167    75        <NA>        gold   yellow     112.0   <NA>
#> 3          R2-D2     96    32        <NA> white, blue      red      33.0   <NA>
#> 4    Darth Vader    202   136        none       white   yellow      41.9   male
#> 5    Leia Organa    150    49       brown       light    brown      19.0 female
#> 6      Owen Lars    178   120 brown, grey       light     blue      52.0   male
#> # ... with 5 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>

camel(lst)

#> $itemA
#> [1] 1 2
#> 
#> $itemB
#> [1] 3 4
#> 

# snake_case
snake(unnamed_vec)

#> [1] "hello_world"       "hello_world"       "rnai_clones"      
#> [4] "worfdb_html_remap" "x123"              "na"               

snake(named_vec)

#>        item_a        item_b 
#> "hello world" "HELLO WORLD" 

snake(df)

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> mazda_rx4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> mazda_rx4_wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> datsun_710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> hornet_4_drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> hornet_sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

snake(df, rownames = FALSE)

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

snake(tbl)

#> # A tibble: 6 x 13
#>             name height  mass  hair_color  skin_color eye_color birth_year
#>            <chr>  <int> <dbl>       <chr>       <chr>     <chr>      <dbl>
#> 1 Luke Skywalker    172    77       blond        fair      blue       19.0
#> 2          C-3PO    167    75        <NA>        gold    yellow      112.0
#> 3          R2-D2     96    32        <NA> white, blue       red       33.0
#> 4    Darth Vader    202   136        none       white    yellow       41.9
#> 5    Leia Organa    150    49       brown       light     brown       19.0
#> 6      Owen Lars    178   120 brown, grey       light      blue       52.0
#> # ... with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>

snake(lst)

#> $item_a
#> [1] 1 2
#> 
#> $item_b
#> [1] 3 4
#> 

# dotted.case
dotted(unnamed_vec)

#> [1] "hello.world"       "hello.world"       "rnai.clones"      
#> [4] "worfdb.html.remap" "x123"              "na"               

dotted(named_vec)

#>        item.a        item.b 
#> "hello world" "HELLO WORLD" 

dotted(df)

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> mazda.rx4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> mazda.rx4.wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> datsun.710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> hornet.4.drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> hornet.sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

dotted(df, rownames = FALSE)

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

dotted(tbl)

#> # A tibble: 6 x 13
#>             name height  mass  hair.color  skin.color eye.color birth.year
#>            <chr>  <int> <dbl>       <chr>       <chr>     <chr>      <dbl>
#> 1 Luke Skywalker    172    77       blond        fair      blue       19.0
#> 2          C-3PO    167    75        <NA>        gold    yellow      112.0
#> 3          R2-D2     96    32        <NA> white, blue       red       33.0
#> 4    Darth Vader    202   136        none       white    yellow       41.9
#> 5    Leia Organa    150    49       brown       light     brown       19.0
#> 6      Owen Lars    178   120 brown, grey       light      blue       52.0
#> # ... with 6 more variables: gender <chr>, homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>

dotted(lst)

#> $item.a
#> [1] 1 2
#> 
#> $item.b
#> [1] 3 4
#>