Parameters

Parameters

Arguments

alpha
numeric(1). Adjusted P value ("alpha") cutoff. If left NULL, will use the cutoff defined in the object.

aspectRatio
integer(1). Aspect ratio.

assay
vector(1). Name or index of count matrix slotted in assays. When passing in a string, the name must be defined in assayNames.

assays
list. Count matrices, which must have matching dimensions. Counts can be passed in as either a dense matrix (matrix) or sparse matrix (sparseMatrix).

bpparam
bpparamClass. BiocParallel parameter specifying the back-end to be used for computations.

We recommend using one of the following:

bpparam.

SerialParam.

MulticoreParam.

censorSamples
character. Specify a subset of samples to censor.

check
logical(1). Perform assert checks.

colData
DataFrame. Metadata describing the assay columns. For bulk RNA-seq, this data describes the samples. For single-cell RNA-seq, this data describes the cells.

color
ScaleDiscrete. Desired ggplot2 color scale. Must supply discrete values. When set NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_color_manual().

To set the discrete color palette globally, use:

options(basejump.color.discrete = ggplot2::scale_color_viridis_d())

counts
matrix. Count matrix. Normalized counts are recommended.

countsAxisLabel
character(1). Counts axis label.

dark
logical(1). Plot against a dark background using the basejump::theme_midnight() ggplot2 theme.

dimsUse
integer. Vector of length 2 that denotes the columns from the reduced dimension matrix to use for centerX and centerY column calculations. Defaults the first and second dimensions.

dir
character(1). Directory path.

direction
character(1). Plot "both", "up", or "down" directions.

ensemblRelease
integer(1). Ensembl release version (e.g. 90). We recommend setting this value if possible, for improved reproducibility. When left unset, the latest release available via AnnotationHub/ensembldb is used. Note that the latest version available can vary, depending on the versions of AnnotationHub and ensembldb in use.

envir
environment. Environment to use for assignment. Defaults to parent.frame(), which will assign into the calling environment.

expression
character(1). Calculation to apply. Uses match.arg() internally and defaults to the first argument in the character vector.

file
character(1). File path.

fill
ggproto/ScaleDiscrete. Desired ggplot2 fill scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_fill_manual().

To set the discrete fill palette globally, use:

options(basejump.fill.discrete = ggplot2::scale_fill_viridis_d())

flip
logical(1). Flip x and y axes. Recommended for plots containing many samples.

gene2symbol
Gene2Symbol. Gene-to-symbol mappings. Must contain geneID and geneName columns. See Gene2Symbol for more information.

genes
character. Gene identifiers. It is considered better practice to input the stable gene identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene symbols (e.g. "TSPN6"), if possible.

genomeBuild
character(1). Ensembl genome build assembly name (e.g. "GRCh38"). If set NULL, defaults to the most recent build available. Note: don't pass in UCSC build IDs (e.g. "hg38").

geom
character(1). Plot type. Uses match.arg() internally and defaults to the first argument in the character vector.

gffFile
character(1). GFF/GTF (General Feature Format) file. Generally, we recommend using a GTF (GFFv2) instead of a GFFv3 file if possible.

headerLevel
integer(1) (1-7). Markdown header level.

i
Indices specifying elements to extract or replace. Indices are numeric or character vectors, empty (missing), or NULL.

For more information:

help(topic = "Extract", package = "base")

inherits
logical(1). Should the enclosing frames of the environment be searched?

interestingGroups
character. Groups of interest that define the samples. If left unset, defaults to sampleName.

label
logical(1). Superimpose sample text labels on the plot.

labelSize
integer(1). Size of the text label.

lfcThreshold
numeric(1). Log fold change ratio (base 2) cutoff threshold. If left NULL, will use the cutoff defined in the object.

legend
logical(1). Show plot legend.

level
character(1). Return ranges as "genes" or "transcripts".

limit
numeric(1). Threshold to denote on the plot, using a dashed line.

max
numeric(1). Recommended maximum value cutoff.

metadata
list. Metadata.

min
numeric(1). Recommended minimum value cutoff.

minCounts
integer(1). Minimum number of counts per gene in the count matrix.

n
integer(1). Number to include.

ntop
integer(1). Number of top genes to label.

object
Object.

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

perMillion
logical(1). Display as counts per million.

perSample
logical(1). Visualize the distributions per sample.

plotlist
list. List containing ggplot objects.

pointAlpha
numeric(1) (0-1). Alpha transparency level. Useful when there many points in the dataset (e.g. single-cell data), and some points can be masked.

pointColor
character(1). Default point color for the plot.

pointScalar
integer(1). Default point size for the plot.

pointsAsNumbers
logical(1). Plot the points as numbers (TRUE) or dots (FALSE).

pointSize
numeric(1). Point size for dots in the plot.

prefilter
logical(1). Apply prefiltering to remove zero count genes.

progress
logical(1). Show progress, using progress bars.

reducedDim
character(1). Name of reduced dimension matrix slotted in reducedDims. Includes TNSE, UMAP, PCA, for example.

return
character(1). Return type. Uses match.arg() internally and defaults to the first argument in the character vector.

rowData
DataFrame. Metadata describing the assay rows, if genomic ranges are not available. Use rowRanges (GRanges) instead, if possible.

rowRanges
GRanges. Genomic ranges (e.g. genome annotations). Metadata describing the assay rows.

sampleMetadataFile
character(1). Sample metadata file path. CSV or TSV is preferred, but Excel worksheets are also supported. Check the documentation for conventions and required columns.

samples
character. Specify a subset of samples to include.

sigPointColor
character. Color names for labeling upregulated and downregulated genes. Also supports a character string for labeling DEGs with the same color, regardless of direction.

sort
logical(1). Resort using sort.

spikeNames
character. Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).

subtitle
character(1). Plot subtitle.

title
character(1). Plot title.

trans
character(1). Name of the axis scale transformation to apply.

For more information:

help(topic = "scale_x_continuous", package = "ggplot2")

transgeneNames
character. Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).

trendline
logical(1). Include trendline on plot.

tx2gene
Tx2Gene. Transcript-to-gene mappings.

url
character(1). Uniform Resource Locator (URL). HTTP or FTP address.

value
Value to assign.

verbose
logical(1). Run the function with verbose messages? It is only recommended to enable this when debugging.

x
Object.

...
Additional arguments.

Value

No value.