Parameters

Parameters

Arguments

object, x
Object.

value
Value to assign.

...
Additional arguments.

i
Indices specifying elements to extract or replace. Indices are numeric or character vectors, empty (missing), or NULL.

For more information:

help(topic = "Extract", package = "base")

assay
vector(1). Name or index of count matrix slotted in assays(). When passing in a string, the name must be defined in assayNames().

colnames
logical(1). Apply to column names.

counts
matrix. Count matrix. Normalized counts are recommended.

dir
character(1). Directory path.

ensemblRelease, release
integer(1). Ensembl release version (e.g. 90). We recommend setting this value if possible, for improved reproducibility. When left unset, the latest release available via AnnotationHub/ensembldb is used. Note that the latest version available can vary, depending on the versions of AnnotationHub and ensembldb in use.

envir
environment. Environment to use for assignment. Defaults to parent.frame(), which will assign into the calling environment.

expression
character(1). Calculation to apply. Uses match.arg() internally and defaults to the first argument in the character vector.

file
character(1). File path.

gene2symbol
Gene2Symbol. Gene-to-symbol mappings. Must contain geneID and geneName columns. See Gene2Symbol for more information.

genes
character. Gene identifiers. It is considered better practice to input the stable gene identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene symbols (e.g. "TSPN6"), if possible.

genomeBuild
character(1). Ensembl genome build assembly name (e.g. "GRCh38"). If set NULL, defaults to the most recent build available. Note: don't pass in UCSC build IDs (e.g. "hg38").

headerLevel
integer(1) (1-7). Markdown header level.

inherits
logical(1). Should the enclosing frames of the environment be searched?

interestingGroups
character. Groups of interest that define the samples. If left unset, defaults to sampleName.

level
character(1). Return as genes or transcripts.

n
integer(1). Number to include.

ntop
integer(1). Number of top genes to label.

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

progress
logical(1). Show progress, using progress bars.

return
character(1). Return type. Uses match.arg() internally and defaults to the first argument in the character vector.

rownames
logical(1). Apply to row names.

samples
character. Specify a subset of samples to include.

sort
logical(1). Resort using sort.

tx2gene
Tx2Gene. Transcript-to-gene mappings.

url
character(1). Uniform Resource Locator (URL). HTTP or FTP address.

assays
list. Count matrices, which must have matching dimensions. Counts can be passed in as either a dense matrix (matrix) or sparse matrix (sparseMatrix).

rowRanges
GRanges. Genomic ranges (e.g. genome annotations). Metadata describing the assay rows.

rowData
DataFrame. Metadata describing the assay rows, if genomic ranges are not available. Use rowRanges (GRanges) instead, if possible.

colData
DataFrame. Metadata describing the assay columns. For bulk RNA-seq, this data describes the samples. For single-cell RNA-seq, this data describes the cells.

metadata
list. Metadata.

transgeneNames
character. Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character. Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).

censorSamples
character. Specify a subset of samples to censor.

gffFile
character(1). GFF/GTF (General Feature Format) file. Generally, we recommend using a GTF (GFFv2) instead of a GFFv3 file if possible.

prefilter
logical(1). Apply prefiltering to remove zero count genes.

sampleMetadataFile
character(1). Sample metadata file path. CSV or TSV is preferred, but Excel worksheets are also supported. Check the documentation for conventions and required columns.

BPPARAM
bpparamClass. BiocParallel parameter to specify the desired processor configuration.
We recommend using one of the following:

bpparam.

SerialParam.

MulticoreParam.

Value

No value.