Prepare SummarizedExperiment

This is a utility wrapper for SummarizedExperiment() that provides automatic subsetting for colData and rowData.

prepareSummarizedExperiment(object, ...)

# S4 method for ANY
prepareSummarizedExperiment(object, colData, rowData,
  metadata = NULL)

# S4 method for data.frame
prepareSummarizedExperiment(object, colData, rowData,
  metadata = NULL)

# S4 method for dgCMatrix
prepareSummarizedExperiment(object, colData, rowData,
  metadata = NULL)

# S4 method for list
prepareSummarizedExperiment(object, colData, rowData,
  metadata = NULL)

# S4 method for matrix
prepareSummarizedExperiment(object, colData, rowData,
  metadata = NULL)

# S4 method for SimpleList
prepareSummarizedExperiment(object, colData, rowData,
  metadata = NULL)

Arguments

object
Object supporting dimensions (base::dim()), or a list containing valid objects. For NGS experiments, a counts matrix is recommended, and can be passed in either dense (matrix) or sparse (dgCMatrix, dgTMatrix) format. Multiple matrices can be supplied as a list, as long as they all have the same dimensions. List object can be supplied as either class list or SimpleList.

...
Additional arguments.

colData
Object describing assay matrix columns. Must support base::dim().

rowData
Object describing assay matrix rows. Must support base::dim().

metadata
Optional. Metadata list.

Value

SummarizedExperiment.

Details

prepareSummarizedExperiment() also provides automatic metadata slotting of multiple useful environment parameters:

date: Today's date.

wd: Working directory.

sessionInfo: R session information.

See also

SummarizedExperiment::SummarizedExperiment().

base::Sys.Date().

base::getwd().

utils::sessionInfo().

Other Data Import and Project Utilities: loadRemoteData, readFileByExtension, readGTF, readYAML, transmit

Examples

mat <- mtcars %>%
    snake(rownames = TRUE) %>%
    .[c("mazda_rx4", "datsun_710"), ] %>%
    .[, c("mpg", "gear")]
colData <- data.frame(
    description = c("Miles per gallon", "Number of gears"),
    abbreviation = c(TRUE, FALSE),
    row.names = colnames(mat))
rowData <- data.frame(
    manufacturer = c("Mazda", "Datsun"),
    model_number = c("RX4", "710"),
    row.names = rownames(mat))
prepareSummarizedExperiment(mat, colData, rowData)

#> Preparing SummarizedExperiment

#> class: SummarizedExperiment 
#> dim: 2 2 
#> metadata(3): date wd sessionInfo
#> assays(1): assay
#> rownames(2): mazda_rx4 datsun_710
#> rowData names(2): manufacturer model_number
#> colnames(2): mpg gear
#> colData names(2): description abbreviation