Read File by Extension

Supports automatic loading of common file extensions:

readFileByExtension(file, ...)

Arguments

file
string. File path.

...
Additional arguments.

Value

Varies, depending on the file extension. Note that data frames are returned as DataFrame class, when applicable.

Details

csv: Comma Separated Values.

gff/gff3/gtf: General Feature Format.

mtx: MatrixMarket sparse matrix.

rda/RData: R Data. Must contain a single object. Doesn't require internal object name to match, like with the loadData() function.

rds: R Data Serialized.

tsv Tab Separated Values.

xlsx: Excel workbook.

yaml/yml: YAML.

Also supports some additional extensions commonly used with the bcbio pipeline:

counts: Counts table.

colnames: Sidecar file containing column names.

rownames: Sidecar file containing row names.

If the file format isn't supported natively (or blacklisted), the rio package will be used as a fallback attempt.

Note

Reading a MatrixMarket ("mtx") file now requires "colnames" and "rownames" sidecar files containing the colnames() and rownames() of the sparse matrix. Legacy support for manual loading of these sidecar files is provided.

See also

Examples

# R Data
x <- readFileByExtension("http://basejump.seq.cloud/rnaseq_counts.rda")

#> Reading rnaseq_counts.rda

glimpse(x)

#>  int [1:1000, 1:12] 0 33 2 18 11 15 93 47 60 25 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:1000] "gene1" "gene2" "gene3" "gene4" ...
#>   ..$ : chr [1:12] "sample1" "sample2" "sample3" "sample4" ...

# Comma Separated Values
x <- readFileByExtension("http://basejump.seq.cloud/mtcars.csv")

#> Reading mtcars.csv

glimpse(x)

#> Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   ..@ rownames       : NULL
#>   ..@ nrows          : int 32
#>   ..@ listData       :List of 12
#>   .. ..$ rowname: chr [1:32] "Mazda RX4" "Mazda RX4 Wag" "Datsun 710" "Hornet 4 Drive" ...
#>   .. ..$ mpg    : num [1:32] 21 21 22.8 21.4 18.7 18.1 14.3 24.4 22.8 19.2 ...
#>   .. ..$ cyl    : int [1:32] 6 6 4 6 8 6 8 4 4 6 ...
#>   .. ..$ disp   : num [1:32] 160 160 108 258 360 ...
#>   .. ..$ hp     : int [1:32] 110 110 93 110 175 105 245 62 95 123 ...
#>   .. ..$ drat   : num [1:32] 3.9 3.9 3.85 3.08 3.15 2.76 3.21 3.69 3.92 3.92 ...
#>   .. ..$ wt     : num [1:32] 2.62 2.88 2.32 3.21 3.44 ...
#>   .. ..$ qsec   : num [1:32] 16.5 17 18.6 19.4 17 ...
#>   .. ..$ vs     : int [1:32] 0 0 1 1 0 1 0 1 1 1 ...
#>   .. ..$ am     : int [1:32] 1 1 1 0 0 0 0 0 0 0 ...
#>   .. ..$ gear   : int [1:32] 4 4 4 3 3 3 3 4 4 4 ...
#>   .. ..$ carb   : int [1:32] 4 4 1 1 2 1 4 2 2 4 ...
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       : list()

# Microsoft Excel Worksheet
x <- readFileByExtension("http://basejump.seq.cloud/mtcars.xlsx")

#> Reading mtcars.xlsx

glimpse(x)

#> Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   ..@ rownames       : NULL
#>   ..@ nrows          : int 6
#>   ..@ listData       :List of 12
#>   .. ..$ rowname: chr [1:6] "Mazda RX4" "Mazda RX4 Wag" "Datsun 710" "Hornet 4 Drive" ...
#>   .. ..$ mpg    : num [1:6] 21 21 22.8 21.4 18.7 18.1
#>   .. ..$ cyl    : num [1:6] 6 6 4 6 8 6
#>   .. ..$ disp   : num [1:6] 160 160 108 258 360 225
#>   .. ..$ hp     : num [1:6] 110 110 93 110 175 105
#>   .. ..$ drat   : num [1:6] 3.9 3.9 3.85 3.08 3.15 2.76
#>   .. ..$ wt     : num [1:6] 2.62 2.88 2.32 3.21 3.44 ...
#>   .. ..$ qsec   : num [1:6] 16.5 17 18.6 19.4 17 ...
#>   .. ..$ vs     : num [1:6] 0 0 1 1 0 1
#>   .. ..$ am     : num [1:6] 1 1 1 0 0 0
#>   .. ..$ gear   : num [1:6] 4 4 4 3 3 3
#>   .. ..$ carb   : num [1:6] 4 4 1 1 2 1
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       : list()

# bcbio Counts Table
x <- readFileByExtension("http://basejump.seq.cloud/example.counts")

#> Reading example.counts

#> Parsed with column specification:
#> cols(
#>   id = col_character(),
#>   sample01 = col_integer(),
#>   sample02 = col_integer(),
#>   sample03 = col_integer(),
#>   sample04 = col_integer(),
#>   sample05 = col_integer(),
#>   sample06 = col_integer(),
#>   sample07 = col_integer(),
#>   sample08 = col_integer(),
#>   sample09 = col_integer(),
#>   sample10 = col_integer(),
#>   sample11 = col_integer(),
#>   sample12 = col_integer()
#> )

glimpse(x)

#>  int [1:9, 1:12] 0 0 259 0 3 0 0 0 0 0 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:9] "ENSMUSG00000102693" "ENSMUSG00000064842" "ENSMUSG00000051951" "ENSMUSG00000102851" ...
#>   ..$ : chr [1:12] "sample01" "sample02" "sample03" "sample04" ...